### 2018-11-14 #263 @ThierryO
git2rdata, a companion package for git2r
Store dataframes as plain text files along with metadata to ensure that attributes like factor levels are maintained. The dataframes are optimized in order to minimize both file size and diffs, making it useful in combination with version control.

```
Package: git2rdata
Title: Store and Retrieve Data.frames in a Git Repository
Version: 0.0.1.9000
  person(
    "Thierry", "Onkelinx", role = c("aut", "cre"),
    email = "thierry.onkelinx@inbo.be",
  person(
    "Research Institute for Nature and Forest",
    role = c("cph", "fnd"), email = "info@inbo.be"))
Description: Make versioning of data.frame easy and efficient using git repositories.
  assertthat,
  git2r (>= 0.23.0),
  readr
Collate:
    'write_vc.R'
    'auto_commit.R'
    'clean_data_path.R'
    'git2rdata-package.R'
    'meta.R'
    'read_vc.R'
    'recent_commit.R'
    'reexport.R'
    'rm_data.R'
```

https://github.com/inbo/git2rdata

[e.g., "data extraction, because the package parses a scientific data file format"]

reproducibility, because it help to store dataframes as plain text files without loss of information, while minimize file sizes and diff.

Anyone who wants to work with medium sized dataframes and have them under version control. This is useful in case of recurrent analysis on growing datasets. E.g. each year new data is added to the data and a new report is created. A snapshot of the raw data can be stored as plain text files under version control.

The initial idea was to add this functionality into `git2r`. After a [discussion](https://github.com/ropensci/git2r/pull/366) with the maintainer, we decide to create a separate package.  We use functions from `read_r` and improve them by storing the metadata.

- [X] includes documentation with examples for all functions. _examples are available in the vignette_

We use `read_vc()` rather than `vc_read()` because it matches `read.table()`, `read_csv()`, ...

@jennybc @stewid
