### 2018-11-20 #263 @noamross
### Editor checks:

---

#### Editor comments

Thanks for your submission @ThierryO! This is an interesting package and it passes checks with nothing but this from `goodpractice::gp()`:

```
  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.
```

I'm seeking reviewers for it now.  FYI, because of the thorny nature of the problem your tackling - data versioning workflows - I anticipate a lot of discussion around high-level concepts and architecture.  I'm going to look for reviewers with a lot of experience on these workflow issues, and encourage them to start that discussion before checking off all the low-level boxes so as to have an efficient and productive review.
---

Reviewers:
Due date:
