### 2018-11-26 2c5385d
bugfix in documentation
