### 2018-11-26 571acc9
add Zenodo DOI to description and README
