+++ DESCRIPTION
@@ -26,7 +26,7 @@ Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 6.1.1
-URL: https://github.com/inbo/git2rdata
+URL: https://github.com/inbo/git2rdata, https://doi.org/10.5281/zenodo.1485309
 BugReports: https://github.com/inbo/git2rdata/issues
 Collate:
     'write_vc.R'
