### 2018-11-28 baeb9ed
bugfix: meta() converts dates into integers
