### 2018-11-28 #263 @noamross
Reviewers assigned: @jennybc and @brodieG.  At reviewers' request, final review date is extended to deal with travel and holidays.  However, I request that reviewers provide some comment on high-level workflow/architecture issues that I think are likely to be the main points of sooner, before before a more detailed look at the code.  Roughly:

-  Is the expected workflow well communicated and implemented in the documentation and vignette?
-  How does the workflow and model of the package work with other frameworks for versioning and reproducibility? Are concepts compatible and separable?
- Is the package API (relatively) intuitive and work well with the workflow?
