### 2018-11-29 #263 @ThierryO
Thanks for assigning the reviewers. I'm looking forward for their comments. There were some minor changes to the package. The latest version is available at inbo/git2rdata#5
