### 2018-11-29 #263 @brodieG
Hi Thierry.  This looks like an interesting package to review; looking forward to it.

One small request: do you mind creating tags/releases that we can work off of so there is no issue of the code changing under our feet as we are reviewing it?  Ideally this would also be accompanied by a unique package version in DESCRIPTION.  This will ensure we're all exactly on the same page about what version of the code is being reviewed.  Thank you.
