### 2018-11-29 #263 @brodieG
Sounds good, I will do all of my work of of [v0.0.1.9000](https://github.com/inbo/git2rdata/tree/v0.0.1.9000).
