### 2018-12-03 2f281d3
read_vc: appropriate error when root is missing
