### 2018-12-04 ddac853
read_vc: set "." as the default root
Thanks @ThierryO for the code suggestion

Co-Authored-By: florisvdh <floris.vanderhaeghe@inbo.be>
