### 2018-12-04 63eb18a
read_vc: show root default in documentation
