### 2018-12-04 e0c582a
write_vc: '.' as default root
