### 2018-12-04 f0327f1
unit test for write_vc with default root
