+++ tests/testthat/test_a_basics.R
@@ -1,6 +1,11 @@
 context("write_vc() and read_vc() on a file system")
-expect_error(write_vc(), "'root' is missing")
 expect_error(write_vc(root = 1), "a 'root' of class numeric is not supported")
+output <- write_vc(x = test_data,
+                   file = "test.txt",
+                   sorting = "test_Date")
+expect_true(all(file.exists(git2rdata:::clean_data_path(".", "test"))))
+if (file.exists("test.tsv")) file.remove("test.tsv")
+if (file.exists("test.yml")) file.remove("test.yml")
 root <- tempfile(pattern = "git2rdata-basic")
 dir.create(root)
 expect_false(any(file.exists(git2rdata:::clean_data_path(root, "test"))))
