### 2018-12-04 fe9ee42
unit test for read_vc with default root
