+++ tests/testthat/test_a_basics.R
@@ -4,6 +4,11 @@ output <- write_vc(x = test_data,
                    file = "test.txt",
                    sorting = "test_Date")
 expect_true(all(file.exists(git2rdata:::clean_data_path(".", "test"))))
+expect_equal(
+  stored <- read_vc(file = "test.xls"),
+  sorted_test_data,
+  check.attributes = FALSE
+)
 if (file.exists("test.tsv")) file.remove("test.tsv")
 if (file.exists("test.yml")) file.remove("test.yml")
 root <- tempfile(pattern = "git2rdata-basic")
