### 2018-12-04 3eb0b52
set default root of rm_data() to '.'
