### 2018-12-04 8365068
unit test for rm_data with default root
