+++ tests/testthat/test_b_rm_data.R
@@ -37,3 +37,14 @@ expect_identical(
   list.files(root, recursive = TRUE),
   current[-grep("^.*/.*", current)]
 )
+
+write_vc(test_data, file = "test", sorting = "test_Date")
+current <- list.files(".", recursive = TRUE)
+expect_identical(
+  rm_data(type = "both", recursive = FALSE),
+  c("test.tsv", "test.yml")
+)
+expect_identical(
+  list.files(".", recursive = TRUE),
+  current[grep("^.*/.*", current)]
+)
