### 2018-12-04 80610ea
extra unit tests for numeric root argument
