### 2018-12-04 984348b
update documentation on root parameter
