+++ README.md
@@ -1,4 +1,4 @@
-# The `git2rdata` package
+# The `git2rdata` package <img src="man/figures/logo.png" align="right" alt="" width="120" />
 
 [![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
 [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
