### 2018-12-22 #263 @ThierryO
Thanks you @brodieG for the elaborate review. Here are my idea's on the topics you raised.

# Editors questions

## Q1

### Vignette

Adding an example where the data is actually used is a good suggestion. I have presented our workflow as a [poster](https://pureportal.inbo.be/portal/files/14910298/onkelinx_2018_isec.pdf) at the International Statistical Ecology Conference [ISEC 2018](http://www.isec2018.org/home). Please note that this poster predates the conception of `git2rdata`. A precursor of the functionality was then available in `n2khelper`.

Tracking analysis code is outside the scope of `git2rdata`. However, I do that in our [`n2kanalysis`](https://github.com/inbo/n2kanalysis) package. This package uses the output of `git2rdata::recent_commit()` as an entry point for the data. The code version is documented as a list of used packages and their version (include commit hash for github package). `n2kanalysis` has even a file and status fingerprints in each object which makes is possible to check the integrity of the object.

If we assume that the code to create and store the data.frames is all stored in a package `foo`, then `gitr2data::auto_commit()` can be used to document this. The commit message will read "Automated commit from 'foo'" and adds the full list of loaded packages and their versions at the moment of the commit. This is of course not tamper proof, but that is not the goal of `git2rdata`.

### Documentation

If required, I can add some toy examples to the `man` documentation. How extensive should these examples be?

## Q2

The initial idea of the package was to read and write data.frames into a git repository. Hence the strong link with Git and `git2r`. Then I had the idea that simply writing data.frames without version control (or to a different version control system than Git) might be relevant as well for other users.

Currently `write_vc()` returns hashes because I use these hashes in our workflow (they are used to calculate the file fingerprints). The file names are returned as the names of the hashes. So the examples with `git2r::add()` and `svnr::commit()` are currently possible when you write them as below. I agree that it is slightly more cumbersome. Returning a names vector with file paths and adding the hashes as names is an option.

```r
git2r::add(path = names(write_vc(df, 'df')))
svnr::commit(names(write_cv(df, 'df')))
```

The current version has only methods depending on the `root` argument, making S3 a relevant option. Although it would require that `root` is the first and mandatory object. S4 allows to write methods for different arguments. E.g. we could think of methods for `x` objects that are something different than a data.frame (e.g. a SpatialPointsDataFrame).

## Q3

### rm_data

- A check on unversioned files when applying `rm_data()` on a git repository is a good idea. Maybe something similar like the `override` argument of `write_vc()`?
- Splitting `rm_data()` over two functions is a great suggestion.

### Possible improvement

I'm not sure if there is much benefit of creating a new class. We often just want to write the object to a file / vcs and don't bother about any intermediate formats.

The `all.equal()` example is rather verbose because I want to stress the fact that `write_vc()` alters the order of the rows and the columns. I'm not sure that the user would need an `all.equal()` method for `git2rdata`. This equality should be warranted by the package itself (and is through unit tests).

# Additional comments

## More Workflow Comments

There already is a [PR](https://github.com/inbo/git2rdata/pull/6) which implements `root = "."` as default. Another option would be to use `rprojroot::find_root()` to find a suitable root. A sensible order of defaults would be `is_git_root`, `is_svn_root`, `is_r_package`, `is_rstudio_project`.

`file = deparse(substitute(x))` is problematic as it becomes `file = "."` when piping is used, as demonstrated in the example below. Furthermore, it assumes that people use informative names for their object (which they seldom do). I prefer to force the user to think about naming the file.

```r
library(dplyr)
test <- function(x, file = deparse(substitute(x))) {
  message(file)
}
test(iris)
iris %>%
  test()
```

## Implications of Versioning Data in Git

A separate vignette with guidelines and caveat on using `git2rdata` on a Git repository would be useful indeed. I'd like to postpone this. This allows me to try it on larger data sets and see were it breaks.

## Rationale

In my line of work, I don't have any control over the raw data source. And hence cannot guarantee that I can reconstruct the raw data. I've been repeating the same analyses every year on several data sources. At least two of them had a major overhaul of the database structure during the past five years. And there is no version control within the database itself.

In the past we published results which were questioned by one of our stakeholders. They disagreed with the results and feared that the results would lead to a detrimental change in policy. The simple fact that we were able state that both the data and the code was placed under version control instantly increased our trustworthiness. After a walk through of the data and the code, they now agree on the methodology and hence the results.

So the focus is not on "the analysis of what changed in the data", but rather on "what data was used in this version of the analysis". Therefore `git2rdata` is just a tool in a fully reproducible analytically pipeline, not the full pipeline.

## Research Of Comparable Solutions

I'll have a look at these suggestions over the next weeks.

## Storage format

Storing each object in a dedicated folder is an interesting suggestion. Another benefit of this, is that we could allow the user to specify "file" extensions. E.g. `write_vc(x, file = "test.txt")` would store the data in `test.txt/data.tsv` and `test.txt/metadata.yml`.

## Dependencies

An [earlier version](https://github.com/inbo/git2rdata/tree/8e0624d848810f39f9a6dfa6c5a3ae1df24d5290) of `git2rdata` relied on `write.table()` and `read.table()`. The downside of that is, that storing characters was somewhat limited: `NA`, `\t` and `\n` were not allowed when storing them unquoted.

Personally, I like `tibble` over `data.frame` because of the superior `print()` method. Adding only the `tibble` dependency would also add at least `fansi`, `fillar` and `utf8`. In case you omit the dependencies in 'Suggested:' it will also required `cli`, `crayon` and `rlang`.

While thinking on adding `tibble`, `readr` can into view. I agree that it again adds several dependencies: `BH`, `hms` and `pkgconfig` (plus `R6` and `Rcpp` which are only dependencies of Suggested dependencies of `git2rdata`).

However, I see several gains in using `readr`

1. `read_vc()` returns a `tibble` (agreed, rather a matter of taste than a critical improvement)
1. `read_tsv()` is faster than `read.table()` as claimed by [the `readr` README](https://github.com/tidyverse/readr), [Efficient R programming](https://csgillespie.github.io/efficientR/5-3-importing-data.html), [this blog](https://analyticstraining.com/if-youre-a-data-analyst-you-should-read-this-review-of-hadleys-readr-0-1-0-right-now/), ...
1. `write_tsv()` handles NA and special characters efficiently by adding quotes only when really needed. And hence code that I don't have to write.
1. `read_tsv()` allows to specify the column type when reading the data.

IMHO, these gains outweigh the extra dependencies.
