### 2019-01-14 #263 @noamross
While waiting for the second review (which I think might take until after `rstudio::conf`), some other quick responses:

#### Examples

> If required, I can add some toy examples to the man documentation. How extensive should these examples be?

There's a lot of flexibility in this, but generally, these should be enough to demonstrate minimal usage/syntax.  I usually make sure to have one example with only required arguments and examples that together include all optional arguments, and that print/show examples of the output value.

#### Dependencies

We outline some [low-dependency options for tibbles in the development guide](https://ropensci.github.io/dev_guide/building.html#package-dependencies), but if the author can justify the trade-offs for dependencies we defer to their choice.
