+++ R/write_vc.R
@@ -40,13 +40,12 @@ setMethod(
     x, file, root, sorting, override = FALSE, optimize = TRUE, ...
   ){
     if (missing(root)) {
-      return(write_vc(x = x,
-                      file = file,
-                      root = ".",
-                      sorting = sorting,
-                      override = FALSE,
-                      optimize = TRUE,
-                      ...))
+      return(
+        write_vc(
+          x = x, file = file, root = ".", sorting = sorting,
+          override = override, optimize = optimize, ...
+        )
+      )
     }
     stop("a 'root' of class ", class(root), " is not supported")
   }
