+++ man/git2rdata-package.Rd
@@ -22,6 +22,7 @@ Useful links:
 
 Other contributors:
 \itemize{
+  \item Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (0000-0002-6378-6229) [contributor]
   \item Research Institute for Nature and Forest \email{info@inbo.be} [copyright holder, funder]
 }
 
