+++ DESCRIPTION
@@ -8,6 +8,10 @@ Authors@R: c(
     email = "thierry.onkelinx@inbo.be", 
     comment = c(ORCID = "0000-0001-8804-4216")),
   person(
+    "Floris", "Vanderhaeghe", role = "ctb", 
+    email = "floris.vanderhaeghe@inbo.be", 
+    comment = c(ORCID = "0000-0002-6378-6229")),
+  person(
     "Research Institute for Nature and Forest",
     role = c("cph", "fnd"), email = "info@inbo.be"))
 Description: Make versioning of data.frame easy and efficient using git repositories.
