### 2019-02-01 inbo/git2rdata#9 @ThierryO
split rm_data() in two functions
As suggested in https://github.com/ropensci/software-review/issues/263#issuecomment-448822999
