### 2019-02-01 inbo/git2rdata#10 @peterdesmet
Move CODE_OF_CONDUCT.md to .github
Suggestion. Nicely bundles these github recognized files together + pkgdown still finds it there.
