### 2019-02-01 inbo/git2rdata#10 @peterdesmet
Note that link in README would have to be updated if you do.
