### 2019-02-01 fa81357
meta() uses S3 instead of S4 (see #8)
