+++ man/meta.Rd
@@ -3,33 +3,9 @@
 \docType{methods}
 \name{meta}
 \alias{meta}
-\alias{meta,character-method}
-\alias{meta,integer-method}
-\alias{meta,numeric-method}
-\alias{meta,factor-method}
-\alias{meta,logical-method}
-\alias{meta,complex-method}
-\alias{meta,POSIXct-method}
-\alias{meta,Date-method}
 \title{Optimize a vector for storage as plain text and add meta data}
 \usage{
 meta(x, optimize = TRUE)
-
-\S4method{meta}{character}(x, optimize = TRUE)
-
-\S4method{meta}{integer}(x, optimize = TRUE)
-
-\S4method{meta}{numeric}(x, optimize = TRUE)
-
-\S4method{meta}{factor}(x, optimize = TRUE)
-
-\S4method{meta}{logical}(x, optimize = TRUE)
-
-\S4method{meta}{complex}(x, optimize = TRUE)
-
-\S4method{meta}{POSIXct}(x, optimize = TRUE)
-
-\S4method{meta}{Date}(x, optimize = TRUE)
 }
 \arguments{
 \item{x}{the vector}
