+++ NAMESPACE
@@ -1,12 +1,20 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(meta,Date)
+S3method(meta,POSIXct)
+S3method(meta,character)
+S3method(meta,complex)
+S3method(meta,factor)
+S3method(meta,integer)
+S3method(meta,logical)
+S3method(meta,numeric)
 export(commit)
+export(meta)
 export(pull)
 export(push)
 export(repository)
 export(status)
 exportMethods(auto_commit)
-exportMethods(meta)
 exportMethods(read_vc)
 exportMethods(recent_commit)
 exportMethods(rm_data)
