### 2019-02-01 0c352d5
meta.character() escapes NA, ", \t and \n
