### 2019-02-01 inbo/git2rdata#11 @ThierryO
Add examples to all helpfiles
As suggested in https://github.com/ropensci/software-review/issues/263#issuecomment-448822999
