### 2019-02-01 bfdea98
Add examples for meta() (#11)
