### 2019-02-01 b886fac
move code of conduct to .github (closes #10)
