+++ README.md
@@ -63,4 +63,4 @@ git2rdata
 
 ## Contributor Code of Conduct
 
-Please note that the 'git2rdata' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+Please note that the 'git2rdata' project is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
