### 2019-02-01 9f5d214
use write.table() instead of readr::write_tsv() (see #7)
