+++ NAMESPACE
@@ -36,7 +36,7 @@ importFrom(methods,setGeneric)
 importFrom(methods,setMethod)
 importFrom(methods,setOldClass)
 importFrom(readr,read_tsv)
-importFrom(readr,write_tsv)
 importFrom(utils,head)
 importFrom(utils,sessionInfo)
 importFrom(utils,tail)
+importFrom(utils,write.table)
