### 2019-02-01 80246da
use read.table() instead of readr::read_tsv() (closes #7)
