+++ NAMESPACE
@@ -35,8 +35,8 @@ importFrom(git2r,workdir)
 importFrom(methods,setGeneric)
 importFrom(methods,setMethod)
 importFrom(methods,setOldClass)
-importFrom(readr,read_tsv)
 importFrom(utils,head)
+importFrom(utils,read.table)
 importFrom(utils,sessionInfo)
 importFrom(utils,tail)
 importFrom(utils,write.table)
