+++ DESCRIPTION
@@ -19,8 +19,7 @@ Depends: R (>= 3.4.0)
 Imports:
   assertthat,
   git2r (>= 0.23.0),
-  methods,
-  readr
+  methods
 Suggests:
   knitr,
   rmarkdown,
