### 2019-02-01 5847f24
convert auto_commit() from S3 to S4 (see #8)
