+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(auto_commit,default)
+S3method(auto_commit,git_repository)
 S3method(meta,Date)
 S3method(meta,POSIXct)
 S3method(meta,character)
@@ -8,13 +10,13 @@ S3method(meta,factor)
 S3method(meta,integer)
 S3method(meta,logical)
 S3method(meta,numeric)
+export(auto_commit)
 export(commit)
 export(meta)
 export(pull)
 export(push)
 export(repository)
 export(status)
-exportMethods(auto_commit)
 exportMethods(read_vc)
 exportMethods(recent_commit)
 exportMethods(rm_data)
