### 2019-02-01 fe369b2
convert read_vc() from S4 to S3 (see #8)
