+++ R/read_vc.R
@@ -2,169 +2,149 @@
 #' @inheritParams write_vc
 #' @return The \code{data.frame}
 #' @rdname read_vc
-#' @exportMethod read_vc
-#' @docType methods
+#' @export
 #' @family storage
-#' @importFrom methods setGeneric
-setGeneric(
-  name = "read_vc",
-  def = function(file, root = "."){
-    standardGeneric("read_vc") # nocov
-  }
-)
+read_vc <- function(file, root) {
+  UseMethod("read_vc", root)
+}
 
-#' @rdname read_vc
-#' @importFrom methods setMethod
-setMethod(
-  f = "read_vc",
-  signature = signature(root = "ANY"),
-  definition = function(file, root){
-    if (missing(root)) {
-      return(read_vc(file = file, root = "."))
-    }
-    stop("a 'root' of class ", class(root), " is not supported")
+#' @export
+read_vc.default <- function(file, root) {
+  if (missing(root)) {
+    return(read_vc(file = file, root = "."))
   }
-)
+  stop("a 'root' of class ", class(root), " is not supported")
+}
 
-#' @rdname read_vc
-#' @importFrom methods setMethod
+#' @export
 #' @importFrom assertthat assert_that is.string
 #' @importFrom utils head read.table
-setMethod(
-  f = "read_vc",
-  signature = signature(root = "character"),
-  definition = function(file, root){
-    assert_that(is.string(file))
-    assert_that(is.string(root))
-    root <- normalizePath(root, winslash = "/", mustWork = TRUE)
+read_vc.character <- function(file, root) {
+  assert_that(is.string(file))
+  assert_that(is.string(root))
+  root <- normalizePath(root, winslash = "/", mustWork = TRUE)
 
-    file <- clean_data_path(root = root, file = file)
+  file <- clean_data_path(root = root, file = file)
+  assert_that(
+    all(file.exists(file)),
+    msg = "raw file and/or meta file missing"
+  )
+
+  meta_data <- readLines(file["meta_file"])
+  start_quote <- grep("^        - \"", meta_data)
+  if (length(start_quote) > 0) {
+    end_quote <- grep("^( {8}- \".*|(?! {8}- ).*)\"$", meta_data, perl = TRUE)
     assert_that(
-      all(file.exists(file)),
-      msg = "raw file and/or meta file missing"
+      length(start_quote) == length(end_quote),
+      all(start_quote <= end_quote),
+      all(head(end_quote, -1) < tail(start_quote, -1)),
+      msg = "Mismatching quotes in metadata"
     )
-
-    meta_data <- readLines(file["meta_file"])
-    start_quote <- grep("^        - \"", meta_data)
-    if (length(start_quote) > 0) {
-      end_quote <- grep("^( {8}- \".*|(?! {8}- ).*)\"$", meta_data, perl = TRUE)
-      assert_that(
-        length(start_quote) == length(end_quote),
-        all(start_quote <= end_quote),
-        all(head(end_quote, -1) < tail(start_quote, -1)),
-        msg = "Mismatching quotes in metadata"
+    for (i in rev(seq_along(start_quote))) {
+      meta_data <- c(
+        meta_data[seq_len(start_quote[i] - 1)],
+        paste(meta_data[start_quote[i]:end_quote[i]], collapse = "\n"),
+        meta_data[(end_quote[i] + 1):length(meta_data)]
       )
-      for (i in rev(seq_along(start_quote))) {
-        meta_data <- c(
-          meta_data[seq_len(start_quote[i] - 1)],
-          paste(meta_data[start_quote[i]:end_quote[i]], collapse = "\n"),
-          meta_data[(end_quote[i] + 1):length(meta_data)]
-        )
-      }
-      meta_data
     }
[…]
