+++ man/read_vc.Rd
@@ -1,20 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/read_vc.R
-\docType{methods}
 \name{read_vc}
 \alias{read_vc}
-\alias{read_vc,ANY,ANY-method}
-\alias{read_vc,ANY,character-method}
-\alias{read_vc,ANY,git_repository-method}
 \title{Read a \code{data.frame} from a repository}
 \usage{
-read_vc(file, root = ".")
-
-\S4method{read_vc}{ANY,ANY}(file, root = ".")
-
-\S4method{read_vc}{ANY,character}(file, root = ".")
-
-\S4method{read_vc}{ANY,git_repository}(file, root = ".")
+read_vc(file, root)
 }
 \arguments{
 \item{file}{the name of the file without file extension. Can include a
