+++ NAMESPACE
@@ -10,14 +10,17 @@ S3method(meta,factor)
 S3method(meta,integer)
 S3method(meta,logical)
 S3method(meta,numeric)
+S3method(read_vc,character)
+S3method(read_vc,default)
+S3method(read_vc,git_repository)
 export(auto_commit)
 export(commit)
 export(meta)
 export(pull)
 export(push)
+export(read_vc)
 export(repository)
 export(status)
-exportMethods(read_vc)
 exportMethods(recent_commit)
 exportMethods(rm_data)
 exportMethods(write_vc)
