### 2019-02-01 e26ddab
convert recent_commit() from S4 to S3 (see #8)
