+++ man/recent_commit.Rd
@@ -1,15 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/recent_commit.R
-\docType{methods}
 \name{recent_commit}
 \alias{recent_commit}
-\alias{recent_commit,ANY,git_repository-method}
 \title{Most recent file change
 Retrieve the most recent commit in which a file or data object existed.}
 \usage{
 recent_commit(file, root, data = FALSE)
-
-\S4method{recent_commit}{ANY,git_repository}(file, root, data = FALSE)
 }
 \arguments{
 \item{file}{the name of the file without file extension. Can include a
