+++ NAMESPACE
@@ -13,15 +13,16 @@ S3method(meta,numeric)
 S3method(read_vc,character)
 S3method(read_vc,default)
 S3method(read_vc,git_repository)
+S3method(recent_commit,git_repository)
 export(auto_commit)
 export(commit)
 export(meta)
 export(pull)
 export(push)
 export(read_vc)
+export(recent_commit)
 export(repository)
 export(status)
-exportMethods(recent_commit)
 exportMethods(rm_data)
 exportMethods(write_vc)
 importFrom(assertthat,assert_that)
