### 2019-02-01 93ef621
convert rm_data() from S4 to S3 (see #8)
