+++ R/rm_data.R
@@ -7,116 +7,96 @@
 #' @return returns invisibily a vector of removed files names. The paths are
 #' relative to `root`.
 #' @inheritParams write_vc
-#' @rdname rm_data
-#' @exportMethod rm_data
-#' @docType methods
+#' @export
 #' @family storage
-#' @importFrom methods setGeneric
-setGeneric(
-  name = "rm_data",
-  def = function(
-    root = ".", path = NULL, type = c("tsv", "yml", "both"), recursive = TRUE,
-    ...
-  ){
-    standardGeneric("rm_data") # nocov
-  }
-)
+rm_data <- function(
+  root, path = NULL, type = c("tsv", "yml", "both"), recursive = TRUE, ...
+){
+  UseMethod("rm_data", root)
+}
 
-#' @rdname rm_data
-#' @importFrom methods setMethod
-setMethod(
-  f = "rm_data",
-  signature = signature(root = "ANY"),
-  definition = function(
-    root, path = NULL, type = c("tsv", "yml", "both"), recursive = TRUE, ...
-  ){
-    if (missing(root)) {
-      return(
-        rm_data(
-          root = ".", path = path, type = type, recursive = recursive, ...
-        )
+#' @export
+rm_data.default <- function(
+  root, path = NULL, type = c("tsv", "yml", "both"), recursive = TRUE, ...
+){
+  if (missing(root)) {
+    return(
+      rm_data(
+        root = ".", path = path, type = type, recursive = recursive, ...
       )
-    }
-    stop("a 'root' of class ", class(root), " is not supported")
+    )
   }
-)
+  stop("a 'root' of class ", class(root), " is not supported")
+}
 
-#' @rdname rm_data
-#' @importFrom methods setMethod
+#' @export
 #' @importFrom assertthat assert_that is.flag
-setMethod(
-  f = "rm_data",
-  signature = signature(root = "character"),
-  definition = function(
-    root, path = NULL, type = c("tsv", "yml", "both"), recursive = TRUE, ...
-  ){
-    assert_that(is.string(root))
-    root <- normalizePath(root, winslash = "/", mustWork = TRUE)
-    assert_that(is.string(path))
-    path <- file.path(root, path)
-    path <- normalizePath(path, winslash = "/", mustWork = FALSE)
-    if (!dir.exists(path)) {
-      return(invisible(NULL))
-    }
-    type <- match.arg(type)
-    assert_that(is.flag(recursive))
-
-    if (type == "tsv") {
-      to_do <- list.files(
-        path = path,
-        pattern = "\\.tsv$",
-        recursive = recursive,
-        full.names = TRUE
-      )
-    } else if (type == "both") {
-      to_do <- list.files(
-        path = path,
-        pattern = "\\.(tsv|yml)$",
-        recursive = recursive,
-        full.names = TRUE
-      )
-    } else {
-      to_do <- list.files(
-        path = path,
-        pattern = "\\.yml$",
-        recursive = recursive,
-        full.names = TRUE
-      )
-      keep <- list.files(
-        path = path,
-        pattern = "\\.tsv$",
-        recursive = recursive,
-        full.names = TRUE
-      )
-      keep <- gsub("\\.tsv$", ".yml", keep)
-      to_do <- to_do[!to_do %in% keep]
-    }
-    file.remove(to_do)
-    to_do <- gsub(paste0("^", root, "/"), "", to_do)
-
-    return(invisible(to_do))
+rm_data.character <- function(
+  root, path = NULL, type = c("tsv", "yml", "both"), recursive = TRUE, ...
+){
+  assert_that(is.string(root))
[…]
