+++ man/rm_data.Rd
@@ -1,23 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rm_data.R
-\docType{methods}
 \name{rm_data}
 \alias{rm_data}
-\alias{rm_data,ANY-method}
-\alias{rm_data,character-method}
-\alias{rm_data,git_repository-method}
+\alias{rm_data.git_repository}
 \title{Remove data files}
 \usage{
-rm_data(root = ".", path = NULL, type = c("tsv", "yml", "both"),
+rm_data(root, path = NULL, type = c("tsv", "yml", "both"),
   recursive = TRUE, ...)
 
-\S4method{rm_data}{ANY}(root = ".", path = NULL, type = c("tsv",
-  "yml", "both"), recursive = TRUE, ...)
-
-\S4method{rm_data}{character}(root = ".", path = NULL,
-  type = c("tsv", "yml", "both"), recursive = TRUE, ...)
-
-\S4method{rm_data}{git_repository}(root, path = NULL, type = c("tsv",
+\method{rm_data}{git_repository}(root, path = NULL, type = c("tsv",
   "yml", "both"), recursive = TRUE, ..., stage = FALSE)
 }
 \arguments{
@@ -32,7 +23,7 @@ Defaults to the current working directory (".").}
 
 \item{...}{additional parameters used in some methods}
 
-\item{stage}{stage the changes after writing the data. Defaults to FALSE}
+\item{stage}{stage the changes after removing the files. Defaults to FALSE.}
 }
 \value{
 returns invisibily a vector of removed files names. The paths are
