+++ NAMESPACE
@@ -14,6 +14,9 @@ S3method(read_vc,character)
 S3method(read_vc,default)
 S3method(read_vc,git_repository)
 S3method(recent_commit,git_repository)
+S3method(rm_data,character)
+S3method(rm_data,default)
+S3method(rm_data,git_repository)
 export(auto_commit)
 export(commit)
 export(meta)
@@ -22,8 +25,8 @@ export(push)
 export(read_vc)
 export(recent_commit)
 export(repository)
+export(rm_data)
 export(status)
-exportMethods(rm_data)
 exportMethods(write_vc)
 importFrom(assertthat,assert_that)
 importFrom(assertthat,is.flag)
