### 2019-02-01 265e5fe
convert write_vc() from S4 to S3 (closes #8)
