+++ R/write_vc.R
@@ -17,154 +17,138 @@
 #' @inheritParams meta
 #' @return a named vector with the hashes of the files. The names contains the
 #' files with their paths relative to `root`.
-#' @rdname write_vc
-#' @exportMethod write_vc
-#' @docType methods
-#' @importFrom methods setGeneric
+#' @export
 #' @family storage
-setGeneric(
-  name = "write_vc",
-  def = function(
-    x, file, root = ".", sorting, override = FALSE, optimize = TRUE, ...
-  ){
-    standardGeneric("write_vc") # nocov
-  }
-)
+write_vc <- function(
+  x, file, root, sorting, override = FALSE, optimize = TRUE, ...
+) {
+  UseMethod("write_vc", root)
+}
 
-#' @rdname write_vc
-#' @importFrom methods setMethod
-setMethod(
-  f = "write_vc",
-  signature = signature(root = "ANY"),
-  definition = function(
-    x, file, root, sorting, override = FALSE, optimize = TRUE, ...
-  ){
-    if (missing(root)) {
-      return(
-        write_vc(
-          x = x, file = file, root = ".", sorting = sorting,
-          override = override, optimize = optimize, ...
-        )
+#' @export
+write_vc.default <- function(
+  x, file, root, sorting, override = FALSE, optimize = TRUE, ...
+) {
+  if (missing(root)) {
+    return(
+      write_vc(
+        x = x, file = file, root = ".", sorting = sorting, override = override,
+        optimize = optimize, ...
       )
-    }
-    stop("a 'root' of class ", class(root), " is not supported")
+    )
   }
-)
+  stop("a 'root' of class ", class(root), " is not supported")
+}
 
-#' @rdname write_vc
-#' @importFrom methods setMethod
+#' @export
 #' @importFrom assertthat assert_that is.string is.flag
 #' @importFrom utils tail write.table
 #' @importFrom git2r hashfile
-setMethod(
-  f = "write_vc",
-  signature = signature(root = "character"),
-  definition = function(
-    x, file, root, sorting, override = FALSE, optimize = TRUE, ...
-  ){
-    assert_that(inherits(x, "data.frame"))
-    assert_that(is.string(file))
-    assert_that(is.string(root))
-    root <- normalizePath(root, winslash = "/", mustWork = TRUE)
-    if (!missing(sorting)) {
-      assert_that(is.character(sorting))
-      assert_that(
-        length(sorting) >= 1,
-        msg = "at least one variable is required for sorting"
-      )
-      assert_that(
-        all(sorting %in% colnames(x)),
-        msg = "use only variables of 'x' for sorting"
-      )
-    }
-    assert_that(is.flag(override))
-    assert_that(is.flag(optimize))
-
-    file <- clean_data_path(root = root, file = file)
-    if (!file.exists(dirname(file["raw_file"]))) {
-      dir.create(dirname(file["raw_file"]), recursive = TRUE)
-    }
-
-    # prepare metadata
-    raw_data <- as.data.frame(
-      lapply(x, meta, optimize = optimize),
-      stringsAsFactors = FALSE
+write_vc.character <- function(
+  x, file, root, sorting, override = FALSE, optimize = TRUE, ...
+){
+  assert_that(inherits(x, "data.frame"))
+  assert_that(is.string(file))
+  assert_that(is.string(root))
+  root <- normalizePath(root, winslash = "/", mustWork = TRUE)
+  if (!missing(sorting)) {
+    assert_that(is.character(sorting))
+    assert_that(
+      length(sorting) >= 1,
+      msg = "at least one variable is required for sorting"
     )
[…]
