+++ man/write_vc.Rd
@@ -1,23 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/write_vc.R
-\docType{methods}
 \name{write_vc}
 \alias{write_vc}
-\alias{write_vc,ANY,ANY,ANY-method}
-\alias{write_vc,ANY,ANY,character-method}
-\alias{write_vc,ANY,ANY,git_repository-method}
+\alias{write_vc.git_repository}
 \title{Write a \code{data.frame} to a git repository}
 \usage{
-write_vc(x, file, root = ".", sorting, override = FALSE,
-  optimize = TRUE, ...)
+write_vc(x, file, root, sorting, override = FALSE, optimize = TRUE,
+  ...)
 
-\S4method{write_vc}{ANY,ANY,ANY}(x, file, root = ".", sorting,
-  override = FALSE, optimize = TRUE, ...)
-
-\S4method{write_vc}{ANY,ANY,character}(x, file, root = ".", sorting,
-  override = FALSE, optimize = TRUE, ...)
-
-\S4method{write_vc}{ANY,ANY,git_repository}(x, file, root, sorting,
+\method{write_vc}{git_repository}(x, file, root, sorting,
   override = FALSE, optimize = TRUE, ..., stage = FALSE,
   force = FALSE)
 }
