+++ NAMESPACE
@@ -17,6 +17,9 @@ S3method(recent_commit,git_repository)
 S3method(rm_data,character)
 S3method(rm_data,default)
 S3method(rm_data,git_repository)
+S3method(write_vc,character)
+S3method(write_vc,default)
+S3method(write_vc,git_repository)
 export(auto_commit)
 export(commit)
 export(meta)
@@ -27,7 +30,7 @@ export(recent_commit)
 export(repository)
 export(rm_data)
 export(status)
-exportMethods(write_vc)
+export(write_vc)
 importFrom(assertthat,assert_that)
 importFrom(assertthat,is.flag)
 importFrom(assertthat,is.string)
@@ -42,7 +45,6 @@ importFrom(git2r,repository)
 importFrom(git2r,status)
 importFrom(git2r,workdir)
 importFrom(methods,setGeneric)
-importFrom(methods,setMethod)
 importFrom(methods,setOldClass)
 importFrom(utils,head)
 importFrom(utils,read.table)
