### 2019-02-04 f363e90
switch to tempdir when testing root = "."
