### 2019-02-04 b46f288
read_vc() uses the origin, format and timezone stored in the metadata
