+++ NAMESPACE
@@ -46,6 +46,7 @@ importFrom(git2r,status)
 importFrom(git2r,workdir)
 importFrom(methods,setGeneric)
 importFrom(methods,setOldClass)
+importFrom(stats,setNames)
 importFrom(utils,head)
 importFrom(utils,read.table)
 importFrom(utils,sessionInfo)
