### 2019-02-04 01f70a4
write_vc() sort on the original dataframe instead of the version altered by meta()
