### 2019-02-04 5314a10
split rm_data() into prune_tsv() and prune_yml(), closes #9
