+++ R/prune.R
@@ -1,6 +1,6 @@
-#' Prune tsv files
+#' Remove data files
 #'
-#' Removes all tsv files from the `path` when they have an accompanying yml file
+#' Removes all data (`.tsv` files) from the `path` when they have accompanying metadata (`.yml` file). The metadata remains untouched.
 #' @param path the directory in which to clean all the data files
 #' @param recursive remove files in subdirectories too
 #' @return returns invisibily a vector of removed files names. The paths are
@@ -8,14 +8,14 @@
 #' @inheritParams write_vc
 #' @export
 #' @family storage
-prune_tsv <- function(
+rm_data <- function(
   root = ".", path = NULL, recursive = TRUE, ...
 ){
-  UseMethod("prune_tsv", root)
+  UseMethod("rm_data", root)
 }
 
 #' @export
-prune_tsv.default <- function(
+rm_data.default <- function(
   root, path = NULL, recursive = TRUE, ...
 ){
   stop("a 'root' of class ", class(root), " is not supported")
@@ -23,7 +23,7 @@ prune_tsv.default <- function(
 
 #' @export
 #' @importFrom assertthat assert_that is.flag
-prune_tsv.character <- function(
+rm_data.character <- function(
   root = ".", path = NULL, recursive = TRUE, ...
 ){
   assert_that(is.string(root))
@@ -58,40 +58,74 @@ prune_tsv.character <- function(
 
 #' @export
 #' @importFrom assertthat assert_that is.flag
-#' @importFrom git2r add
+#' @importFrom git2r workdir add
 #' @include write_vc.R
 #' @param stage stage the changes after removing the files. Defaults to FALSE.
-#' @rdname prune_tsv
-prune_tsv.git_repository <- function(
-  root, path = NULL, recursive = TRUE, ..., stage = FALSE
+#' @param type which classes of files should be removed. `unmodified` are files in the git history and unchanged since the last commit. `modified` are files in the git history and changed since the last commit. `ignored` refers to file listed in a `.gitignore` file. Selecting `modified` will remove both `unmodified` and `modified` data files. Selecting `ìgnored` will remove `unmodified`, `modified` and `ignored` data files. `all` refers to all visible data files, inclusing `untracked` files. The argument can be abbreviated to the first letter.
+#' @rdname rm_data
+rm_data.git_repository <- function(
+  root, path = NULL, recursive = TRUE, ..., stage = FALSE,
+  type = c("unmodified", "modified", "ignored", "all")
 ){
+  assert_that(is.string(path))
   assert_that(is.flag(stage))
-  removed <- prune_tsv(
-    root = workdir(root), path = path, recursive = recursive, ...
-  )
-  if (stage && !is.null(removed)) {
-    add(repo = root, path = removed)
+  type <- match.arg(type)
+  path <- file.path(workdir(root), path)
+  path <- normalizePath(path, winslash = "/", mustWork = FALSE)
+  if (!dir.exists(path)) {
+    return(invisible(NULL))
   }
-  return(invisible(removed))
+
+  to_do <- list.files(
+    path, pattern = "\\.tsv$", recursive = recursive, full.names = TRUE
+  )
+  yml <- list.files(
+    path, pattern = "\\.yml$", recursive = recursive, full.names = TRUE
+  )
+  yml <- gsub("\\.yml$", ".tsv", yml)
+  to_do <- to_do[to_do %in% yml]
+
+  keep <- unlist(switch(type,
+    unmodified = status(
[…]
