+++ tests/testthat/test_c_git.R
@@ -3,15 +3,18 @@ root <- tempfile(pattern = "git2rdata-")
 dir.create(root)
 root <- git2r::init(root)
 git2r::config(root, user.name = "Alice", user.email = "alice@example.org")
-writeLines("ignore.*", file.path(git2r::workdir(root), ".gitignore"))
+writeLines("ignore.*\nforce.*", file.path(git2r::workdir(root), ".gitignore"))
 git2r::add(root, ".gitignore")
 commit(root, "initial commit")
 untracked <- write_vc(
   test_data, file = "untracked", root = root, sorting = "test_Date"
 )
 expect_equal(
-  status(root),
-  list(staged = list(), unstaged = list(), untracked = names(untracked)),
+  status(root, ignored = TRUE),
+  list(
+    staged = list(), unstaged = list(), untracked = names(untracked),
+    ignored = list()
+  ),
   check.attributes = FALSE
 )
 expect_equal(
@@ -32,8 +35,11 @@ staged <- write_vc(
   test_data, file = "staged", root = root, sorting = "test_Date", stage = TRUE
 )
 expect_equal(
-  status(root),
-  list(staged = names(staged), unstaged = list(), untracked = names(untracked)),
+  status(root, ignored = TRUE),
+  list(
+    staged = names(staged), unstaged = list(), untracked = names(untracked),
+    ignored = list()
+  ),
   check.attributes = FALSE
 )
 expect_equal(
@@ -54,13 +60,12 @@ ignored <- write_vc(
   test_data, file = "ignore", root = root, sorting = "test_Date", stage = TRUE
 )
 expect_equal(
-  status(root),
-  list(staged = names(staged), unstaged = list(), untracked = names(untracked)),
+  status(root, ignored = TRUE),
+  list(
+    staged = names(staged), unstaged = list(), untracked = names(untracked),
+    ignored = names(ignored)
+  ),
   check.attributes = FALSE
-)
-expect_identical(
-  list.files(workdir(root)),
-  c(names(ignored), names(staged), names(untracked))
 )
 expect_equal(
   stored <- read_vc(file = "ignore", root = root),
@@ -76,19 +81,22 @@ for (i in colnames(stored)) {
   )
 }
 
-ignored <-
-  write_vc(test_data, file = "ignore", root = root, stage = TRUE, force = TRUE)
+forced <- write_vc(
+  test_data, file = "forced/force", root = root, sorting = "test_Date",
+  stage = TRUE, force = TRUE
+)
 expect_equal(
-  status(root),
+  status(root, ignored = TRUE),
   list(
-    staged = c(names(ignored), names(staged)),
+    staged = c(names(forced), names(staged)),
     unstaged = list(),
-    untracked = names(untracked)
+    untracked = names(untracked),
+    ignored = names(ignored)
   ),
   check.attributes = FALSE
 )
 expect_equal(
-  stored <- read_vc(file = "ignore", root = root),
+  stored <- read_vc(file = "forced/force", root = root),
   sorted_test_data,
   check.attributes = FALSE
 )
@@ -107,8 +115,11 @@ staged <- write_vc(
   file = "staged", root = root, stage = FALSE
 )
 expect_equal(
-  status(root),
-  list(staged = list(), unstaged = "staged.tsv", untracked = names(untracked)),
+  status(root, ignored = TRUE),
+  list(
+    staged = list(), unstaged = "staged.tsv", untracked = names(untracked),
+    ignored = names(ignored)
+  ),
   check.attributes = FALSE
 )
 expect_equal(
@@ -130,8 +141,11 @@ staged <- write_vc(
[…]

[…]
