+++ tests/testthat/test_b_prune.R
@@ -4,6 +4,7 @@ expect_error(rm_data(root = 1), "a 'root' of class numeric is not supported")
 expect_error(prune_meta(root = 1), "a 'root' of class numeric is not supported")
 
 root <- tempfile(pattern = "git2rdata-")
+root <- normalizePath(root, winslash = "/", mustWork = FALSE)
 expect_error(rm_data(root), root)
 expect_error(prune_meta(root), root)
 dir.create(root)
@@ -36,7 +37,7 @@ expect_identical(list.files(root, recursive = TRUE), current)
 file.remove(file.path(root, "test.tsv"))
 
 old_wd <- getwd()
-setwd(tempdir(check = TRUE))
+setwd(tempdir())
 write_vc(test_data, file = "test", sorting = "test_Date")
 current <- list.files(".", recursive = TRUE, pattern = c("\\.(tsv|yml)$"))
 expect_identical(rm_data(path = ".", recursive = FALSE), "test.tsv")
