### 2019-02-04 dffd93e
add example for auto_commit() (see #11)
