### 2019-02-04 3e9bf03
add missing unit tests
