### 2019-02-04 0ed30db
add examples for rm_data() and prune_tsv() (see #11)
