+++ R/prune.R
@@ -8,6 +8,7 @@
 #' @inheritParams write_vc
 #' @export
 #' @family storage
+#' @template example-prune
 rm_data <- function(
   root = ".", path = NULL, recursive = TRUE, ...
 ){
@@ -119,6 +120,7 @@ rm_data.git_repository <- function(
 #' @inheritParams write_vc
 #' @export
 #' @family storage
+#' @template example-prune
 prune_meta <- function(
   root = ".", path = NULL, recursive = TRUE, ...
 ){
