+++ man/prune_meta.Rd
@@ -29,6 +29,30 @@ relative to \code{root}.
 \description{
 Removes all metadata (\code{.yml} files) from the \code{path} when they don't have accompanying data (\code{.tsv} file).
 }
+\examples{
+# on file system
+root <- tempfile("git2rdata-")
+dir.create(root)
+write_vc(iris[1:6, ], "iris", root, sorting = "Sepal.Length")
+list.files(root, recursive = TRUE)
+rm_data(root, path = ".")
+list.files(root, recursive = TRUE)
+prune_meta(root, path = ".")
+list.files(root, recursive = TRUE)
+
+# on git repo
+repo_path <- tempfile("git2rdata-repo-")
+dir.create(repo_path)
+repo <- git2r::init(repo_path)
+write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
+status(repo)
+auto_commit(package = "my_package", repo)
+status(repo)
+rm_data(repo, path = ".")
+status(repo)
+prune_meta(repo, path = ".")
+status(repo)
+}
 \seealso{
 Other storage: \code{\link{read_vc}},
   \code{\link{rm_data}}, \code{\link{write_vc}}
