### 2019-02-04 5fe5d8c
add examples to write_vc() and read_vc() (see #11)
