+++ R/read_vc.R
@@ -4,6 +4,7 @@
 #' @rdname read_vc
 #' @export
 #' @family storage
+#' @template example-io
 read_vc <- function(file, root = ".") {
   UseMethod("read_vc", root)
 }
