### 2019-02-04 94e4b77
add examples for recent_commit() (closes #11)
