### 2019-02-04 23bf2f8
no need to test root = "." as it is the default
