### 2019-02-04 a8e9965
bugfix: config git repo in examples
