+++ R/recent_commit.R
@@ -13,6 +13,7 @@
 #' repo_path <- tempfile("git2rdata-repo")
 #' dir.create(repo_path)
 #' repo <- git2r::init(repo_path)
+#' git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 #' write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
 #' auto_commit("important analysis", repo)
 #' Sys.sleep(1.1) # required because git doesn't handle subsecond timings
