+++ man-roxygen/example-prune.R
@@ -13,6 +13,7 @@
 #' repo_path <- tempfile("git2rdata-repo-")
 #' dir.create(repo_path)
 #' repo <- git2r::init(repo_path)
+#' git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 #' write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
 #' status(repo)
 #' auto_commit(package = "my_package", repo)
