+++ man/read_vc.Rd
@@ -43,6 +43,7 @@ list.files(root, recursive = TRUE)
 repo_path <- tempfile("git2rdata-repo-")
 dir.create(repo_path)
 repo <- git2r::init(repo_path)
+git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 write_vc(iris[1:6, ], file = "iris", root = repo, sorting = "Sepal.Length")
 status(repo)
 read_vc("iris", repo)
