### 2019-02-05 fc769e5
write_vc() returns the file names using their hashes as names rather than the other way around
