### 2019-02-05 8b0c2e8
read_vc() adds the source file names and their hashes as attributes
