### 2019-02-05 f3b1dbb
don't allow storage of NA string in characters or unoptimized factors
