+++ tests/testthat/test_b_special.R
@@ -8,7 +8,7 @@ ds <- data.frame(
     "a\nb", "a\nb\nc", "\na", "a\n",
     "a\"b", "a\"b\"c", "\"b", "a\"", "\"b\"",
     "a'b", "a'b'c", "'b", "a'", "'b'",
-    "a b c"
+    "a b c", "\"NA\"", "'NA'", NA
   ),
   stringsAsFactors = FALSE
 )
@@ -21,13 +21,33 @@ expect_equal(
   ds[order(ds$a), , drop = FALSE], # nolint
   check.attributes = FALSE
 )
-ds$a <- factor(ds$a)
 expect_is(
-  write_vc(ds, "factor", root, sorting = "a"),
+  write_vc(ds, "character2", root, sorting = "a", optimize = FALSE),
+  "character"
+)
+expect_equal(
+  junk <- read_vc("character2", root),
+  ds[order(ds$a), , drop = FALSE], # nolint
+  check.attributes = FALSE
+)
+z <- rbind(ds, "NA")
+z$a <- factor(z$a)
+expect_is(
+  write_vc(z, "factor", root, sorting = "a"),
   "character"
 )
 expect_equal(
   read_vc("factor", root),
+  z[order(z$a), , drop = FALSE], # nolint
+  check.attributes = FALSE
+)
+ds$a <- factor(ds$a)
+expect_is(
+  write_vc(ds, "factor2", root, sorting = "a", optimize = FALSE),
+  "character"
+)
+expect_equal(
+  read_vc("factor2", root),
   ds[order(ds$a), , drop = FALSE], # nolint
   check.attributes = FALSE
 )
