### 2019-02-06 #263 @jtr13
##### For packages co-submitting to JOSS

- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

- [not reviewed] **Functionality:** Any functional claims of the software been confirmed.

- [not reviewed] **Performance:** Any performance claims of the software been confirmed.

- [not reviewed] **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

- [not reviewed] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

---

#### Note

In the interest of not delaying the review, I was not able to fully test all functionality claims.

#### Documentation comments

##### Statement of need

The first two paragraphs of the vignette provide a much more detailed explanation of the problem that the software is designed to solve than the README. I know there's been a lot of discussion about what should go where and the importance of not having to maintain duplicate text. That said, I am of the opinion that important information like this should be in the README since my sense is that vignettes are often overlooked.  (More so now since `build_vignettes = TRUE` [doesn't work](https://community.rstudio.com/t/vignettes-suddenly-stopped-installing/18391)).

##### Examples

I see that examples were added in the `develop` branch. More comments would be helpful to guide the user through the examples.

##### Community Guidelines

There is no CONTRIBUTING file nor contributing guidelines in the README. `URL`, `BugReports` and `Maintainer` are provided in DESCRIPTION file.

#### Functionality comments

I was not able to perform a thorough review to confirm functional and performance claims.

All code in the vignette worked as expected as did additional toy examples.

The package claims to address three needs:

1) version control of data

2) optimization of data storage

3) meta data storage to preserve data types, factor levels, etc.

Taking these in reverse order: storing meta data  in a `.yml` file has clear and obvious benefits. Losing factor information in particular is a source of frustration for R users and the package provides functions that could be useful to even beginning R users.

The benefits of optimization are not as clearly apparent. It would be helpful to be provided with more context, that is, answers to questions such as: in typical use cases, how much smaller are the text files created by `git2rdata` functions? Is there a time tradeoff?

Finally, with regard to version control, the contribution of `git2rdata`--to the best of my understanding--is to sort rows and columns to minimize diffs. The work of actually maintaining version histories is performed by git, with `git2r` providing the interface to R.

It should be noted though that `git2r` is optional even for using the main `git2rdata` functions: `write_vc` and `read_vc`. Particularly for beginning users, using these functions with the RStudio git interface would be a simpler workflow.

Given this scenario, the package name, `git2rdata` is somewhat misleading and doesn't capture the breadth of contributions that the package makes. In addition, it limits flexibility in making these functions available to other version controls as suggested in other reviews. From the name it sounds as though the package would perform similar functions to `git2r` in a new domain, that is, add functionality for git commands that are not available in `git2r`.  That, however, is not what it does. The vignette title "Storing Dataframes" is more on the mark; perhaps a name along the lines of `storedf` would be a better fit. (On a personal note, I've had a lot of trouble with `git2r` so I wonder if the link might scare off users who would otherwise benefit from this package.)

Some minor points: it's not clear what the benefit is of using `rm_data()` vs. other means to remove files or overwrite them.

If the sorting parameter isn't essential it would be helping to have a way to not use it (`sorting = FALSE`) or have some other default. For example, if all you care about is preserving factor information, it would be nice to have a simple workflow for doing so.
