### 2019-02-11 inbo/git2rdata#12 @ThierryO
Update README
As suggested  by https://github.com/ropensci/software-review/issues/263#issuecomment-460899144

- [x] Elaborate on the rationale of `git2rdata`
- [x] Add a link to the contributing guidelines
- [x] Update installation instructions using `remotes` instead of `devtools` and add option to build vignettes
- [x] Add a link to the vignettes
- [X] add ROpenSci review badge
- [x] move lifecycle badge from "experimental" to "maturing"
- [x] move status badge from "WIP" to "Active"
- [x] add short summary of the vignettes to the README
- [x] add some rough guidelines about maximal size of the data
