### 2019-02-11 #263 @ThierryO
Thank you for your feedback, @jtr13.

## Documentation comments

I'll update the README (see https://github.com/inbo/git2rdata/issues/12). A CONTRIBUTING.md file is present in the `.github` folder. A link will be added to the README.

I'll add a some comments to the examples in the help files.

## General comments

I'll probably separate (and elaborate) the current vignette into several vignettes.

- Basic intro with the concept, without using version control
- Comparison in terms of efficiency in storage and timing by the optimization
- A workflow storing data using version control outside `git2rdata`
- A workflow storing data using version control from within `git2rdata`
- A worflow _analysing_ stored data using version control using `git2rdata`

Making the sorting optional is OK when not storing the files under version control. However, then it is up to the user to use the sorting when writing to a file path and doing the version control outside `git2rdata`. Which is something they will probably forget and end up with potentially large diffs...

The `rm_data()` and `prune_meta()` are useful in an automated workflow where changes in the raw data can lead to the removal of data sets. First apply `rm_data()` to remove **all** tsv files. Then write all the data using `write_vc()`. This is (re)create the tsv files which are still needed. Since all tsv were removed at the start, the ones we don't need any more are no longer present. The final step is to use `prune_meta()` to remove leftover yml files with no matching tsv file.

## Import, suggest or remove `git2r`?

The functionality of `git2rdata` seems a bit of overkill when **not** using version control. Storing dataframes as binary files using `save()` or `saveRDS()` is probably more efficient in such case. Git is a very popular version control system so most likely the one used in combination with `git2rdata`. Hence we think it will be more easily found when the name refers to git.

The `git2r` dependency is not strictly necessary when not using git. Or in an _ad hoc_ workflow where the user manually writes data, stages the required files and creates the commits.

Another use case is an automated one with recurrent data storage and subsequent analysis. The user writes a script that queries a raw data source and stores the resulting dataframes under version control. Under such scenario, the user will most likely want to commit all dataframes written by the function. The user can use `git2r::add(git2rdata::write_vc())`, or the more convient and readable `git2rdata::write_vc(stage = TRUE)`. The user probably also want to commit the changes at the end of script. One options is `git2r::commit()` potentially in combination with `git2r::push()`. Another option is to use `git2rdata::auto_commit()` or `git2rdata::auto_commit(push = TRUE)`. The main benefit of `auto_commit` is that it creates a standardised commit message listing the loaded R packages and there versions, thus documenting the code that was used to create the commit. Suppose that the script is stored into a custom package. Then the message from `auto_commit()` is sufficient to document which _code_ was used to create the commit. The code itself should document _why_ and _how_ the data is created.

One way to document the data used in subsequent analysis is to use the commit hash of the data that is checked out. In our use case, we have several different dataframes stored under version control. The same statistical model is fit to each of them. Not every dataframe will be updated in the next commit of the data. Knowing that the data underlying a model didn't change, eliminates the need to refit that specific model, thus saving time and resources. This is were `recent_commit()` is useful as it lists the latest commit in which a file was changed.

To summarise:

- `auto_commit()` and `recent_commit()` are useful functionality in some workflows to warrant the dependency of `git2r`, at least as a _Suggests_ package.
- Using `git2r` as an _Imports_ package has the benefit of cleaner code. About one third of the current code base depends on `git2r`. It also allows to re-export core `git2r` functions like `repository()`, `status()`, `commit()`, `pull()` and `push()`. This reduces the need for the user to load both `git2r` and `git2rdata` for the basic functionality in an automated workflow.
