### 2019-02-11 64126b0
add a link to CONTRIBUTION.md to the README #12
