### 2019-02-11 4506785
update the installation instructions in the README
