+++ vignettes/efficiency.Rmd
@@ -1,0 +1,142 @@
+---
+title: "Efficiency in terms of storage and time"
+author: "Thierry Onkelinx"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Efficiency in terms of storage and time}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+library(knitr)
+opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+par(cex = 0.5)
+```
+
+## Introduction
+
+This vignette compares storage and retrieval of data by `git2rdata` with other standard R functionality. We consider `write.table()` and `read.table()` for data stored in a plain text format. `saveRDS()` and `readRDS()` use a compressed binary format.
+
+In order to get some meaningful results, we will use the `nassCDS` dataset from the [DAAG](https://www.rdocumentation.org/packages/DAAG/versions/1.22/topics/nassCDS) package. We'll avoid the dependency on the package by directly downloading the data.
+
+```{r}
+airbag <- read.csv(
+  "https://vincentarelbundock.github.io/Rdatasets/csv/DAAG/nassCDS.csv"
+)
+airbag$dead <- airbag$dead == "dead"
+airbag$airbag <- airbag$airbag == "airbag"
+airbag$seatbelt <- airbag$seatbelt == "belted"
+airbag$dvcat <- as.ordered(airbag$dvcat)
+str(airbag)
+```
+
+```{r}
+library(git2rdata)
+root <- tempfile("git2rdata-efficient")
+dir.create(root)
+```
+
+## Data storage when writing to a file system
+
+We start by writing the dataset as is with `write.table()`, `saveRDS()`, `write_vc()` and `write_vc()` without storage optimisation. Note that `write_vc()` uses optimisation by default. Since `write_vc()` creates two files for each data set, we take their combinated file size into account.
+
+```{r}
+write.table(airbag, file.path(root, "base_R.tsv"), sep = "\t")
+base_size <- file.size(file.path(root, "base_R.tsv"))
+
+saveRDS(airbag, file.path(root, "base_R.rds"))
+rds_size <- file.size(file.path(root, "base_R.rds"))
+
+fn <- write_vc(airbag, "airbag_optimize", root, sorting = "X")
+optim_size <- sum(file.size(file.path(root, fn)))
+
+fn <- write_vc(airbag, "airbag_verbose", root, sorting = "X", optimize = FALSE)
+verbose_size <- sum(file.size(file.path(root, fn)))
+```
+
+Since the data is highly compressable, `saveRDS()` yields the smallest file at the cost of having a binary file format. Both `write_vc()` formats yield smaller files than `write.table()`. Partly because `write_vc()` doesn't store row names and only uses quotes when needed. The difference between the optimized and verbose version of `write_vc()` is, in this case, solely due to the way factors are stored in the data (tsv) file. The optimized version stores the indices of the factor whereas the verbose version stores the levels. For example: `airbag$dvcat` has 5 levels with fairly short levels (on average 5 character), however storing the index requires only 1 character. Resulting in more compact files.
+
+```{r echo = FALSE}
+kable(
+  data.frame(
+    method = c("saveRDS()", "write_vc(), optimized", "write_vc(), verbose", 
+               "write.table()"),
+    file_size = c(rds_size, optim_size, verbose_size, base_size) / 2 ^ 10,
+    relative = c(rds_size, optim_size, verbose_size, base_size) / base_size
+  ),
[…]
