### 2019-02-13 #263 @noamross
Thanks @brodieG and @jtr13 for your reviews, and @ThierryO for your thoughtful comments.  I think this discussion has helped clarify the workflows that the package does and does not try to address, and the extended documentation you discuss would go a long way in helping users understand this.

A couple of observations: dependency weight and the return of tibbles seems to be one of the most contentious issues among our author and reviewer communities.  We mostly defer to authors on this.  In general, we find dependency _weight_ to be a bigger issue for developers than users.   We also find its helpful to give users choices as well as sensible dependencies. An package-level option such as `options(git2rdata.tibbles=TRUE)` may be useful.

The tightness of the recommended workflows with version control, and utility of `git2r` in facilitating this, even if individual functions can be separated into their vcs and non-vcs components, seems to me to warrant having `git2r` Imported so as to make combined convenience functions available to the users.  (I find automated versioning of a data set changing due to updates and corrections a compelling example case, and suggest an example incorporating continuous integration may be useful.)

Finally, while we won't require a package name change, I agree with both @ThierryO that having `git` in the name is useful for discovery, and @jtr13 that the current name with `git2r`, (and the description above, "a companion package for git2r"), does make it seem that it primarily extends the domain of the `git2r` package. Something that evokes the workflow may be more informative.
