+++ vignettes/efficiency.Rmd
@@ -7,11 +7,14 @@ vignette: >
   %\VignetteIndexEntry{Efficiency in terms of storage and time}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+  %\VignetteDepends{git2r}
+  %\VignetteDepends{microbenchmark}
 ---
 
 ```{r setup, include = FALSE}
 library(knitr)
 opts_chunk$set(
+  fig.height = 4, fig.width = 6,
   collapse = TRUE,
   comment = "#>"
 )
@@ -24,7 +27,7 @@ This vignette compares storage and retrieval of data by `git2rdata` with other s
 
 In order to get some meaningful results, we will use the `nassCDS` dataset from the [DAAG](https://www.rdocumentation.org/packages/DAAG/versions/1.22/topics/nassCDS) package. We'll avoid the dependency on the package by directly downloading the data.
 
-```{r}
+```{r download_data, eval = system.file("efficiency", "airbag.rds", package = "git2rdata") == ""}
 airbag <- read.csv(
   "https://vincentarelbundock.github.io/Rdatasets/csv/DAAG/nassCDS.csv"
 )
@@ -32,39 +35,54 @@ airbag$dead <- airbag$dead == "dead"
 airbag$airbag <- airbag$airbag == "airbag"
 airbag$seatbelt <- airbag$seatbelt == "belted"
 airbag$dvcat <- as.ordered(airbag$dvcat)
+```
+
+```{r load_data, echo = FALSE}
+if (system.file("efficiency", "airbag.rds", package = "git2rdata") == "") {
+  saveRDS(airbag, file.path("..", "inst", "efficiency", "airbag.rds"))
+} else {
+  airbag <- readRDS(
+    system.file("efficiency", "airbag.rds", package = "git2rdata")
+  )
+}
+```
+
+```{r data_structure}
 str(airbag)
 ```
 
-```{r}
+## Data storage 
+
+### On a file system
+
+We start by writing the dataset as is with `write.table()`, `saveRDS()`, `write_vc()` and `write_vc()` without storage optimisation. Note that `write_vc()` uses optimisation by default. Since `write_vc()` creates two files for each data set, we take their combinated file size into account.
+
+```{r set_tmp_dir}
 library(git2rdata)
 root <- tempfile("git2rdata-efficient")
 dir.create(root)
 ```
 
-## Data storage when writing to a file system
-
-We start by writing the dataset as is with `write.table()`, `saveRDS()`, `write_vc()` and `write_vc()` without storage optimisation. Note that `write_vc()` uses optimisation by default. Since `write_vc()` creates two files for each data set, we take their combinated file size into account.
-
-```{r}
+```{r file_size}
 write.table(airbag, file.path(root, "base_R.tsv"), sep = "\t")
 base_size <- file.size(file.path(root, "base_R.tsv"))
 
 saveRDS(airbag, file.path(root, "base_R.rds"))
 rds_size <- file.size(file.path(root, "base_R.rds"))
 
-fn <- write_vc(airbag, "airbag_optimize", root, sorting = "X")
+fn <- write_vc(airbag, "airbag_optimise", root, sorting = "X")
 optim_size <- sum(file.size(file.path(root, fn)))
 
-fn <- write_vc(airbag, "airbag_verbose", root, sorting = "X", optimize = FALSE)
+fn <- write_vc(airbag, "airbag_verbose", root, sorting = "X", optimise = FALSE)
 verbose_size <- sum(file.size(file.path(root, fn)))
 ```
 
[…]
