+++ DESCRIPTION
@@ -21,6 +21,7 @@ Imports:
   methods
 Suggests:
   knitr,
+  microbenchmark,
   rmarkdown,
   testthat
 License: GPL-3
