### 2019-02-15 68ab772
make read_vc() faster
