### 2019-02-15 47cce5f
add timing on reading data. closes #2
