### 2019-02-15 #263 @brodieG
@ThierryO I see that you have made several changes.  Do you have a new version we should work off of now (if you do, a tag of that version would be great)?  Do you have other changes in the works we should wait for you to complete before we move to the detailed review?  @noamross what's the time frame for the detailed review?  One thing I'm wondering is whether it makes sense to wait for the vignettes outlined in @ThierryO most recent response.
