### 2019-02-15 #263 @noamross
@brodieG I suggest waiting for @ThierryO to complete current changes including vignettes before giving the package another look.
