### 2019-02-18 6a28937
add effeciency vignette to pkgdown website
