### 2019-02-18 1497105
display timings as boxplots rather than density plots
