+++ vignettes/efficiency.Rmd
@@ -248,13 +248,9 @@ if (system.file("efficiency", "file_timings.rds", package = "git2rdata") == "") 
 
 `write_vc()` takes 75% to 100% more time than `write.table()` because it needs to prepare the metadata and sort the observations and variables. When overwriting existing data, the new data is checked against the existing metadata.
 
-```{r plot_file_timings, echo = FALSE, fig.cap = "Density of the write timings for the different methods. black: saveRDS(), red: write.table(), green: write_vc(), optimised, blue: write_vc(), verbose."}
-plot(density(mb$time[mb$expr == "saveRDS"]), xlim = c(90, 420), 
-     xlab = "Time (in milliseconds)", ylab = "Density", 
-     main = "Density of timings")
-lines(density(mb$time[mb$expr == "write.table"]), col = "red")
-lines(density(mb$time[mb$expr == "write_vc.optim"]), col = "green")
-lines(density(mb$time[mb$expr == "write_vc.verbose"]), col = "blue")
+```{r plot_file_timings, echo = FALSE, fig.cap = "Boxplot of the write timings for the different methods."}
+mb$expr <- reorder(mb$expr, mb$time, FUN = median)
+boxplot(time ~ expr, data = mb, ylab = "Time (in milliseconds)")
 ```
 
 ### Reading data
@@ -282,11 +278,8 @@ if (system.file("efficiency", "read_timings.rds", package = "git2rdata") == "") 
 
 The timings on reading the data is is more extreme story. Reading the binary format takes about 8% of the time needed to read the standard plain text format using `read.table()`. `read_vc()` takes about 70% of the time needed by `read.table()`, which at first seems strange because `read_vc()` calls `read.table()` to read the files and has some extra work to convert the variables to the correct data type. The main difference is that `read_vc()` knows the required data type _a priori_ and passes this info to `read.table()`. Otherwise, `read.table()` has to guess the correct data type from the file.
 
-```{r plot_read_timings, echo = FALSE, fig.cap = "Density of the read timings for the different methods. black: readRDS(), red: read.table(), green: read_vc(), optimised, blue: read_vc(), verbose."}
-plot(density(mb$time[mb$expr == "readRDS"]), xlim = c(0, 250), 
-     xlab = "Time (in milliseconds)", ylab = "Density", 
-     main = "Density of timings")
-lines(density(mb$time[mb$expr == "read.table"]), col = "red")
-lines(density(mb$time[mb$expr == "read_vc.optim"]), col = "green")
-lines(density(mb$time[mb$expr == "read_vc.verbose"]), col = "blue")
+```{r plot_read_timings, echo = FALSE, fig.cap = "Boxplots for the read timings for the different methods."}
+mb$expr <- reorder(mb$expr, mb$time, FUN = median)
+boxplot(time ~ expr, data = mb, ylim = c(0, max(mb$time)), 
+        ylab = "Time (in milliseconds)")
 ```
