### 2019-02-18 #263 @ThierryO
@noamross : what is the policy on suggested packages? I'm thinking about using ggplot2 for plots in the vignettes.
