### 2019-02-18 d150a1b
bugfix: use correct argument name in vignette
