+++ vignettes/efficiency.Rmd
@@ -55,7 +55,7 @@ str(airbag)
 
 ### On a file system
 
-We start by writing the dataset as is with `write.table()`, `saveRDS()`, `write_vc()` and `write_vc()` without storage optimisation. Note that `write_vc()` uses optimisation by default. Since `write_vc()` creates two files for each data set, we take their combinated file size into account.
+We start by writing the dataset as is with `write.table()`, `saveRDS()`, `write_vc()` and `write_vc()` without storage optimization. Note that `write_vc()` uses optimization by default. Since `write_vc()` creates two files for each data set, we take their combinated file size into account.
 
 ```{r set_tmp_dir}
 library(git2rdata)
@@ -70,19 +70,19 @@ base_size <- file.size(file.path(root, "base_R.tsv"))
 saveRDS(airbag, file.path(root, "base_R.rds"))
 rds_size <- file.size(file.path(root, "base_R.rds"))
 
-fn <- write_vc(airbag, "airbag_optimise", root, sorting = "X")
+fn <- write_vc(airbag, "airbag_optimize", root, sorting = "X")
 optim_size <- sum(file.size(file.path(root, fn)))
 
-fn <- write_vc(airbag, "airbag_verbose", root, sorting = "X", optimise = FALSE)
+fn <- write_vc(airbag, "airbag_verbose", root, sorting = "X", optimize = FALSE)
 verbose_size <- sum(file.size(file.path(root, fn)))
 ```
 
-Since the data is highly compressable, `saveRDS()` yields the smallest file at the cost of having a binary file format. Both `write_vc()` formats yield smaller files than `write.table()`. Partly because `write_vc()` doesn't store row names and only uses quotes when needed. The difference between the optimised and verbose version of `write_vc()` is, in this case, solely due to the way factors are stored in the data (tsv) file. The optimised version stores the indices of the factor whereas the verbose version stores the levels. For example: `airbag$dvcat` has 5 levels with fairly short levels (on average 5 character), however storing the index requires only 1 character. Resulting in more compact files.
+Since the data is highly compressable, `saveRDS()` yields the smallest file at the cost of having a binary file format. Both `write_vc()` formats yield smaller files than `write.table()`. Partly because `write_vc()` doesn't store row names and only uses quotes when needed. The difference between the optimized and verbose version of `write_vc()` is, in this case, solely due to the way factors are stored in the data (tsv) file. The optimized version stores the indices of the factor whereas the verbose version stores the levels. For example: `airbag$dvcat` has 5 levels with fairly short levels (on average 5 character), however storing the index requires only 1 character. Resulting in more compact files.
 
 ```{r table_file_size, echo = FALSE}
 kable(
   data.frame(
-    method = c("saveRDS()", "write_vc(), optimised", "write_vc(), verbose", 
+    method = c("saveRDS()", "write_vc(), optimized", "write_vc(), verbose", 
                "write.table()"),
     file_size = c(rds_size, optim_size, verbose_size, base_size) / 2 ^ 10,
     relative = c(rds_size, optim_size, verbose_size, base_size) / base_size
@@ -92,18 +92,18 @@ kable(
 )
 ```
 
[…]
