### 2019-02-18 d81336a
only define the optimize argument in the methods where it is relevant
