### 2019-02-18 f6b08de
meta.character() and meta.factor() gain an na argument
