+++ tests/testthat/test_a_basics.R
@@ -1,8 +1,11 @@
 context("write_vc() and read_vc() on a file system")
-expect_error(meta("NA"), "NA is not allowed as character value")
+expect_error(meta("NA"), "one of the strings matches the NA string")
+expect_error(meta("NA", na = "abc def"), "na contains whitespace characters")
+expect_error(meta("NA", na = "abc\tdef"), "na contains whitespace characters")
+expect_error(meta("NA", na = "abc\ndef"), "na contains whitespace characters")
 expect_error(
   meta(factor("NA"), optimize = FALSE),
-  "NA is not allowed as factor level"
+  "one of the levels matches the NA string"
 )
 expect_error(write_vc(root = 1), "a 'root' of class numeric is not supported")
 expect_error(read_vc(root = 1), "a 'root' of class numeric is not supported")
