+++ NAMESPACE
@@ -35,9 +35,11 @@ export(repository)
 export(rm_data)
 export(status)
 export(write_vc)
+importFrom(assertthat,"on_failure<-")
 importFrom(assertthat,assert_that)
 importFrom(assertthat,is.flag)
 importFrom(assertthat,is.string)
+importFrom(assertthat,noNA)
 importFrom(git2r,add)
 importFrom(git2r,commit)
 importFrom(git2r,hashfile)
