### 2019-02-19 8da9a6e
write_vc() can use a user specified NA string
