### 2019-02-20 a4bc9a8
change the storage of metadata in preparation of #13
