+++ tests/testthat/test_a_basics.R
@@ -12,10 +12,6 @@ expect_error(read_vc(root = 1), "a 'root' of class numeric is not supported")
 root <- tempfile(pattern = "git2rdata-basic")
 dir.create(root)
 expect_false(any(file.exists(git2rdata:::clean_data_path(root, "test"))))
-expect_error(
-  write_vc(x = test_data, file = "test.txt", root = root),
-  "new metadata requires 'sorting'"
-)
 expect_is(
   output <- write_vc(
     x = test_data, file = "test.txt", root = root, sorting = "test_Date"
@@ -44,14 +40,19 @@ expect_identical(
 )
 expect_error(
   write_vc(x = test_data, file = "test", root = root, optimize = FALSE),
-  "old data was stored optimized"
+  "new data is verbose, whereas old data was optimized"
+)
+expect_warning(
+  write_vc(x = test_data, file = "test", root = root, optimize = FALSE,
+           strict = FALSE),
+  "new data is verbose, whereas old data was optimized"
 )
 expect_error(
   write_vc(
     x = test_data[, colnames(test_data) != "test_Date"],
     file = "test", root = root
   ),
-  "new data lacks old sorting variable"
+  "all sorting variables must be available"
 )
 
 expect_false(any(file.exists(git2rdata:::clean_data_path(root, "a/verbose"))))
@@ -79,7 +80,7 @@ for (i in colnames(stored)) {
 }
 expect_error(
   write_vc(x = test_data, file = "a/verbose", root = root),
-  "old data was stored verbose"
+  "new data is optimized, whereas old data was verbose"
 )
 
 expect_is(
@@ -108,12 +109,8 @@ expect_error(
   "sorting is not a character vector"
 )
 expect_error(
-  write_vc(test_data, file = "error", root = root, sorting = character()),
-  "at least one variable is required for sorting"
-)
-expect_error(
   write_vc(test_data, file = "error", root = root, sorting = "junk"),
-  "use only variables of 'x' for sorting"
+  "all sorting variables must be available"
 )
 expect_false(any(file.exists(git2rdata:::clean_data_path(root, "sorting"))))
 expect_warning(
@@ -126,43 +123,31 @@ expect_true(all(file.exists(git2rdata:::clean_data_path(root, "sorting"))))
 test_changed <- test_data
 test_changed$junk <- test_changed$test_character
 expect_error(
-  write_vc(
+  suppressWarnings(write_vc(
     test_changed, file = "sorting", root = root, sorting = "test_factor"
-  ),
-  "old data has different number of variables"
+  )),
+  "new data has a different number of variables"
 )
 test_changed$test_character <- NULL
 expect_error(
-  write_vc(
+  suppressWarnings(write_vc(
     test_changed, file = "sorting", root = root, sorting = "test_factor"
-  ),
-  "old data has different variables"
+  )),
+  "new variables: junk\ndeleted variables: test_character"
 )
 test_changed <- test_data
 test_changed$test_character <- factor(test_changed$test_character)
 expect_error(
-  write_vc(
+  suppressWarnings(write_vc(
     test_changed, file = "sorting", root = root, sorting = "test_factor"
+  )),
+  "change in class: test_character from character to factor"
+)
+expect_error(
+  suppressWarnings(
+    write_vc(test_data, file = "sorting", root = root, sorting = "test_logical")
   ),
-  "old data has different variable types or sorting"
-)
-expect_error(
-  write_vc(
-    test_data, file = "sorting", root = root, sorting = "test_logical"
-  ),
-  "old data has different variable types or sorting"
-)
-
-yml <- file.path(root, "sorting.yml")
[…]
