+++ R/write_vc.R
@@ -10,9 +10,9 @@
 #' @param sorting a vector of column names defining which columns to use for
 #' sorting \code{x} and in what order to use them. Only required when writing
 #' new metadata.
-#' @param override Ignore existing meta data. This is required when new
-#' variables are added or variables are deleted. Setting this to TRUE can
-#' potentially lead to large diffs. Defaults to FALSE.
+#' @param strict What to do when the metadata changes. `strict = FALSE` will
+#' overwrite the data with a warning listing the changes, `strict = TRUE` will
+#' return an error and leave the data as is. Default to `TRUE`
 #' @param ... additional parameters used in some methods
 #' @inheritParams meta
 #' @inheritParams utils::write.table
@@ -22,7 +22,7 @@
 #' @family storage
 #' @template example-io
 write_vc <- function(
-  x, file, root = ".", sorting, override = FALSE, optimize = TRUE, na = "NA",
+  x, file, root = ".", sorting, strict = TRUE, optimize = TRUE, na = "NA",
   ...
 ) {
   UseMethod("write_vc", root)
@@ -30,113 +30,58 @@ write_vc <- function(
 
 #' @export
 write_vc.default <- function(
-  x, file, root, sorting, override = FALSE, optimize = TRUE, na = "NA", ...
+  x, file, root, sorting, strict = TRUE, optimize = TRUE, na = "NA", ...
 ) {
   stop("a 'root' of class ", class(root), " is not supported")
 }
 
 #' @export
 #' @importFrom assertthat assert_that is.string is.flag
-#' @importFrom utils tail write.table
+#' @importFrom yaml read_yaml write_yaml
+#' @importFrom utils write.table
 #' @importFrom git2r hashfile
 write_vc.character <- function(
-  x, file, root = ".", sorting, override = FALSE, optimize = TRUE, na = "NA",
+  x, file, root = ".", sorting, strict = TRUE, optimize = TRUE, na = "NA",
   ...
 ){
   assert_that(
     inherits(x, "data.frame"), is.string(file), is.string(root),  is.string(na),
-    noNA(na), no_whitespace(na), is.flag(override), is.flag(optimize)
+    noNA(na), no_whitespace(na), is.flag(strict), is.flag(optimize)
   )
   root <- normalizePath(root, winslash = "/", mustWork = TRUE)
-  if (!missing(sorting)) {
-    assert_that(is.character(sorting))
-    assert_that(
-      length(sorting) >= 1,
-      msg = "at least one variable is required for sorting"
-    )
-    assert_that(
-      all(sorting %in% colnames(x)),
-      msg = "use only variables of 'x' for sorting"
-    )
-  }
-
   file <- clean_data_path(root = root, file = file)
   if (!file.exists(dirname(file["raw_file"]))) {
     dir.create(dirname(file["raw_file"]), recursive = TRUE)
   }
 
-  # prepare metadata
-  raw_data <- as.data.frame(
-    lapply(x, meta, optimize = optimize, na = na),
-    stringsAsFactors = FALSE
-  )
-  metadata <- paste(
-    colnames(x),
-    vapply(raw_data, attr, "", which = "meta"),
-    sep = ":\n"
-  )
-  names(metadata) <- colnames(x)
-  if (override || !file.exists(file["meta_file"])) {
-    # write new metadata
+  if (file.exists(file["meta_file"])) {
+    old <- read_yaml(file["meta_file"])
+    class(old) <- "meta_list"
     if (missing(sorting)) {
-      stop("new metadata requires 'sorting'")
+      sorting <- old[["..generic"]][["sorting"]]
+      if (is.null(sorting)) {
+        sorting <- character(0)
+      }
     }
[…]
