### 2019-02-20 b4a515c
bugfix in comparing metadata between versions
