### 2019-02-21 19a5b84
use vapply instead of sapply
