### 2019-02-21 42c2c0e
add relabel() to change factor labels in the metadata
