+++ NAMESPACE
@@ -22,6 +22,8 @@ S3method(read_vc,character)
 S3method(read_vc,default)
 S3method(read_vc,git_repository)
 S3method(recent_commit,git_repository)
+S3method(relabel,default)
+S3method(relabel,list)
 S3method(rm_data,character)
 S3method(rm_data,default)
 S3method(rm_data,git_repository)
@@ -36,6 +38,7 @@ export(pull)
 export(push)
 export(read_vc)
 export(recent_commit)
+export(relabel)
 export(repository)
 export(rm_data)
 export(status)
