### 2019-02-21 89e6f27
relabel() gains a method for data.frames with changes. closes #13
