+++ NAMESPACE
@@ -22,6 +22,7 @@ S3method(read_vc,character)
 S3method(read_vc,default)
 S3method(read_vc,git_repository)
 S3method(recent_commit,git_repository)
+S3method(relabel,data.frame)
 S3method(relabel,default)
 S3method(relabel,list)
 S3method(rm_data,character)
