### 2019-02-21 e7cc930
update NEWS and bump package version
