+++ man/rm_data.Rd
@@ -59,6 +59,7 @@ status(repo)
 }
 \seealso{
 Other storage: \code{\link{prune_meta}},
-  \code{\link{read_vc}}, \code{\link{write_vc}}
+  \code{\link{read_vc}}, \code{\link{relabel}},
+  \code{\link{write_vc}}
 }
 \concept{storage}
