+++ man/relabel.Rd
@@ -46,3 +46,9 @@ change <- data.frame(
 )
 relabel("relabel", root, change)
 }
+\seealso{
+Other storage: \code{\link{prune_meta}},
+  \code{\link{read_vc}}, \code{\link{rm_data}},
+  \code{\link{write_vc}}
+}
+\concept{storage}
