+++ NEWS.md
@@ -1,3 +1,32 @@
+git2rdata 0.0.2 (2019-02-21)
+============================
+
+### BREAKING CHANGES
+
+  * metadata is added as a list to the objects rather than in YAML format.
+  * the [yaml](https://cran.r-project.org/package=yaml) package is used to store the metadata list in YAML format.
+  * `write_vc()` now used the 'strict' argument instead of 'override'
+  * the functionality `rm_data()` is split into `rm_data()` and `prune_meta()` (#9)
+
+### NEW FEATURES
+
+  * vignette on [efficiency](../articles/efficiency.html) added (#2)
+  * S4 methods are replaced by S3 methods (#8)
+  * optimized factors use stable indices, resulting in smaller diffs when levels are added or removed (#13)
+  * use `relabel()` to alter factor levels without changing their index (#13)
+  * the raw data is written and read by base R functions instead of `readr` functions (#7)
+  * `write_vc()` and `read_vc()` use the current working directory as default root (#6, @florisvdh)
+  * the user can specify a string to code missing values (default = `NA`). This allows the storage of the character string `"NA"`.
+  * `write_vc()` returns a list of issues which potentially result in large diffs.
+
+### Other changes
+
+  * each helpfile contains a working example (#11)
+  * README updated (#12)
+      * `git2rdata` has an hexsticker logo
+      * A DOI is added
+      * The installation instructions uses `remotes` and build the vignettes
+
 git2rdata 0.0.1 (2018-11-12)
 ============================
 
