+++ man/prune_meta.Rd
@@ -56,6 +56,7 @@ status(repo)
 }
 \seealso{
 Other storage: \code{\link{read_vc}},
-  \code{\link{rm_data}}, \code{\link{write_vc}}
+  \code{\link{relabel}}, \code{\link{rm_data}},
+  \code{\link{write_vc}}
 }
 \concept{storage}
