### 2019-02-22 7c3df7d
add vignette on plain text files
