+++ _pkgdown.yml
@@ -5,8 +5,8 @@ navbar:
   - text: Tutorials
     href: articles/index.html
     menu:
-    - text: Introduction
-      href: articles/storing_dataframe.html
+    - text: Storing dataframes as plain text
+      href: articles/plain_text.html
     - text: Efficiency
       href: articles/efficiency.html
   - text: Functions
@@ -16,10 +16,12 @@ navbar:
   right:
   - text: NEWS
     href: news/index.html
-  - icon: fa-twitter fa-lg
-    href: https://twitter.com/INBOVlaanderen
   - icon: fa-github fa-lg
     href: https://github.com/inbo/git2rdata
+  - icon: fa-twitter fa-lg
+    href: https://twitter.com/INBOVlaanderen
+  - icon: fa-facebook fg-lg
+    href: https://www.facebook.com/pg/INBOVlaanderen
 
 reference:
 - title: Storage
