### 2019-02-22 fe4ba89
meta() hides the call on errors
