### 2019-02-22 518650b
errors in write_vc() don't return the class
