### 2019-02-23 e3b29d7
bugfix: write_vc() reuses existing factor indices
thanks to @florisvdh for spotting this in #13
