### 2019-02-23 5344595
relabel() uses the current working directory as default (as was documented) #13
thanks to @florisvdh for spotting this
