### 2019-02-25 #263 @noamross
Sorry for the late reply on this one, @ThierryO.  Suggested packages are are fine, especially for the purpose of displaying outputs in vignettes.
