### 2019-02-25 32600ee
bugfix: relabel works with factor dataframes
