### 2019-02-25 be7235d
use ggplot2 in efficiency vignette
