+++ vignettes/efficiency.Rmd
@@ -1,14 +1,16 @@
 ---
 title: "Efficiency in terms of storage and time"
 author: "Thierry Onkelinx"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
+output: 
+  rmarkdown::html_vignette:
+        fig_caption: yes
 vignette: >
   %\VignetteIndexEntry{Efficiency in terms of storage and time}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
   %\VignetteDepends{git2r}
   %\VignetteDepends{microbenchmark}
+  %\VignetteDepends{ggplot2}
 ---
 
 ```{r setup, include = FALSE}
@@ -18,7 +20,101 @@ opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-par(cex = 0.5)
+library(ggplot2)
+inbo_colours <- c("#959B38", "#729BB7", "#E87837", "#BDDDD7", "#E4E517", 
+                  "#843860", "#C04384", "#C2C444", "#685457")
+theme_inbo <- function(base_size = 12, base_family = "") {
+  rect.bg <- "white"
+  legend.bg <- "white"
+  panel.bg <- "#F3F3F3"
+  panel.grid <- "white"
+  plot.bg <- "white"
+  half_line <- base_size / 2
+  theme(
+    line = element_line(colour = "black", size = 0.5, linetype = 1, 
+                        lineend = "butt"),
+    rect = element_rect(fill = rect.bg, colour = "black", size = 0.5, 
+                        linetype = 1),
+    text = element_text(family = base_family, face = "plain", 
+                        colour = "#843860", size = base_size, hjust = 0.5, 
+                        vjust = 0.5, angle = 0, lineheight = 0.9, 
+                        margin = margin(), debug = FALSE),
+    axis.line = element_blank(),
+    axis.line.x = element_blank(),
+    axis.line.y = element_blank(),
+    axis.text = element_text(size = rel(0.8)),
+    axis.text.x = element_text(margin = margin(t = 0.8 * half_line / 2), 
+                               vjust = 1),
+    axis.text.x.top = NULL,
+    axis.text.y = element_text(margin = margin(r = 0.8 * half_line / 2), 
+                               hjust = 1),
+    axis.text.y.right = NULL,
+    axis.ticks = element_line(),
+    axis.ticks.length = unit(0.15, "cm"),
+    axis.title = element_text(colour = "black"),
+    axis.title.x = element_text(
+      margin = margin(t = 0.8 * half_line, b = 0.8 * half_line / 2)
+    ),
+    axis.title.x.top = NULL,
+    axis.title.y = element_text(
+      margin = margin(r = 0.8 * half_line, l = 0.8 * half_line / 2),
+      angle = 90
+    ),
+    axis.title.y.right = NULL,
+    legend.background = element_rect(colour = NA, fill = legend.bg),
+    legend.key = element_rect(fill = panel.bg, colour = NA),
+    legend.key.size = unit(1.2, "lines"),
+    legend.key.height = NULL,
+    legend.key.width = NULL,
+    legend.margin = NULL,
+    legend.spacing = unit(0.2, "cm"),
+    legend.spacing.x = NULL,
+    legend.spacing.y = NULL,
+    legend.text = element_text(size = rel(0.8)),
+    legend.text.align = NULL,
+    legend.title = element_text(size = rel(0.8), face = "bold", hjust = 0, 
+                                colour = "black"),
+    legend.title.align = NULL,
+    legend.position = "right",
+    legend.direction = NULL,
+    legend.justification = "center",
+    legend.box = NULL,
+    legend.box.margin = margin(half_line, half_line, half_line, half_line),
[…]
