+++ DESCRIPTION
@@ -21,6 +21,7 @@ Imports:
   methods,
   yaml
 Suggests:
+  ggplot2,
   knitr,
   microbenchmark,
   rmarkdown,
