+++ man-roxygen/example-prune.R
@@ -3,11 +3,14 @@
 #' root <- tempfile("git2rdata-")
 #' dir.create(root)
 #' write_vc(iris[1:6, ], "iris", root, sorting = "Sepal.Length")
+#' list_data(root)
 #' list.files(root, recursive = TRUE)
 #' rm_data(root, path = ".")
 #' list.files(root, recursive = TRUE)
+#' list_data(root)
 #' prune_meta(root, path = ".")
 #' list.files(root, recursive = TRUE)
+#' list_data(root)
 #'
 #' # on git repo
 #' repo_path <- tempfile("git2rdata-repo-")
@@ -16,9 +19,12 @@
 #' git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 #' write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
 #' status(repo)
+#' list_data(repo)
 #' auto_commit(package = "my_package", repo)
 #' status(repo)
 #' rm_data(repo, path = ".")
+#' list_data(repo)
 #' status(repo)
 #' prune_meta(repo, path = ".")
+#' list_data(repo)
 #' status(repo)
