+++ man/prune_meta.Rd
@@ -34,11 +34,14 @@ Removes all metadata (\code{.yml} files) from the \code{path} when they don't ha
 root <- tempfile("git2rdata-")
 dir.create(root)
 write_vc(iris[1:6, ], "iris", root, sorting = "Sepal.Length")
+list_data(root)
 list.files(root, recursive = TRUE)
 rm_data(root, path = ".")
 list.files(root, recursive = TRUE)
+list_data(root)
 prune_meta(root, path = ".")
 list.files(root, recursive = TRUE)
+list_data(root)
 
 # on git repo
 repo_path <- tempfile("git2rdata-repo-")
@@ -47,11 +50,14 @@ repo <- git2r::init(repo_path)
 git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
 status(repo)
+list_data(repo)
 auto_commit(package = "my_package", repo)
 status(repo)
 rm_data(repo, path = ".")
+list_data(repo)
 status(repo)
 prune_meta(repo, path = ".")
+list_data(repo)
 status(repo)
 }
 \seealso{
