### 2019-02-25 2c2dba3
bugfix: recent_commit() takes the currently checkout commit into account
