+++ R/recent_commit.R
@@ -1,6 +1,6 @@
 #' Most recent file change
 #'
-#' Retrieve the most recent commit in which a file or data object existed.
+#' Retrieve the most recent commit in which a file or data object was added or updated.
 #' @inheritParams write_vc
 #' @param root The root of a project. Can be a file path or a `git-repository`
 #' @param data does `file` refers to a data object (TRUE) or to a file (FALSE).
@@ -44,16 +44,16 @@ recent_commit <- function(file, root, data = FALSE){
 #' @importFrom assertthat assert_that is.string is.flag
 #' @importFrom git2r odb_blobs
 recent_commit.git_repository <- function(file, root, data = FALSE) {
-  assert_that(is.string(file))
-  assert_that(is.flag(data))
+  assert_that(is.string(file), is.flag(data))
 
-  if (data) {
+  if (isTRUE(data)) {
     file <- clean_data_path(root = ".", file, normalize = FALSE)
   }
   name <- basename(file)
   path <- gsub("^\\./?", "", unique(dirname(file)))
   blobs <- odb_blobs(root)
   blobs <- blobs[blobs$path == path & blobs$name %in% name, ]
+  blobs <- blobs[blobs$when <= as.data.frame(last_commit(root))$when, ]
   blobs <- blobs[blobs$when == max(blobs$when), c("commit", "author", "when")]
   blobs <- unique(blobs)
   if (nrow(blobs) > 1) {
