+++ R/list_data.R
@@ -5,6 +5,7 @@
 #' @export
 #' @template example-prune
 #' @return a character vector is dataframe names, including their relative path
+#' @family storage
 list_data <- function(root = ".", path = ".", recursive = TRUE) {
   UseMethod("list_data", root)
 }
