+++ man/read_vc.Rd
@@ -64,8 +64,8 @@ write_vc(
 status(repo)
 }
 \seealso{
-Other storage: \code{\link{prune_meta}},
-  \code{\link{relabel}}, \code{\link{rm_data}},
-  \code{\link{write_vc}}
+Other storage: \code{\link{list_data}},
+  \code{\link{prune_meta}}, \code{\link{relabel}},
+  \code{\link{rm_data}}, \code{\link{write_vc}}
 }
 \concept{storage}
