+++ man/list_data.Rd
@@ -50,3 +50,9 @@ prune_meta(repo, path = ".")
 list_data(repo)
 status(repo)
 }
+\seealso{
+Other storage: \code{\link{prune_meta}},
+  \code{\link{read_vc}}, \code{\link{relabel}},
+  \code{\link{rm_data}}, \code{\link{write_vc}}
+}
+\concept{storage}
