+++ man/rm_data.Rd
@@ -64,8 +64,8 @@ list_data(repo)
 status(repo)
 }
 \seealso{
-Other storage: \code{\link{prune_meta}},
-  \code{\link{read_vc}}, \code{\link{relabel}},
-  \code{\link{write_vc}}
+Other storage: \code{\link{list_data}},
+  \code{\link{prune_meta}}, \code{\link{read_vc}},
+  \code{\link{relabel}}, \code{\link{write_vc}}
 }
 \concept{storage}
