### 2019-02-25 0b6e442
add vignette with potential workflows
