+++ vignettes/storing_dataframe.Rmd
@@ -1,300 +1,0 @@
----
-title: "Storing dataframes as plain text files suitable for version control"
-author: "Thierry Onkelinx"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Storing dataframes as plain text files suitable for version control}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include = FALSE}
-library(knitr)
-opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-## Concept
-
-### Maintaining variable classes
-
-R has several options to store dataframes as plain text files from R. Base R has `write.table()` and its companions like `write.csv()`. The `readr` package has `write_delim()`, `write_csv()` and `write_tsv()`. Each of them writes a dataframe as a plain text file by converting all variables into characters. After reading the file, the conversion is reversed. However, the distinction between `character` and `factor` is lost in translation. `read.table()` converts by default all strings to factors, `read_csv()` keeps by default all strings as character. The factor levels are another thing which is lost. The `read` function determines the factor levels based on the observed levels in the plain text file. Hence factor levels without observations will disappear. The order of the factor levels is also determined by the available levels in the plain text file, which can be different from the original order.
-
-The `write_vc()` and `read_vc()` functions from `git2rdata` keep track of the class of each variable and, in case of a factor, also of the factor levels and their order. Hence this function pair preserves the information content of the dataframe. The `vc` suffix stands for **v**ersion **c**ontrol as these functions use their full capacity in combination with a version control system.
-
-### Maintaining the order of columns and rows
-
-While `write.table()` and the like write the dataframe as is, `write_vc()` will order both the variables and the rows. Why? Simply because we prefer small and meaningful diffs when storing files under version control. The diffs in most version control systems are row based: a line is marked as a change in the diff as soon as the new line is different from the old line. 
-
-#### Swapping variables
-
-Suppose we have a very simple dataframe as below. Swapping `A` and `B` might not seem a big deal. The information content is the same (assuming there is no information in the order of the variables). The file sizes will be identical. The diff, however, is maximal: each row has a change.
-
-Original version
-```
-A,B
-1,10
-2,11
-3,12
-```
-
-Altered version
-```
-B,A
-10,1
-11,2
-12,3
-```
-
-Diff between original and altered version
-```diff
--A,B
-+B,A
--1,10
-+10,1
--2,11
-+11,2
--3,12
-+12,3
-```
-
-`write_vc()` solves this problem by storing the order of the variables in the metadata (see [Basic usage without version control](#basic-usage)). In case a file is updated, the variables in the new version are ordered as indicated in the metadata from the previous version.
-
-#### Moving a row
-
-Moving a row will not affect the file size nor the information content either. Again, we assume that the order of the observation has no information. The diff of moving a single row will result in the deletion at the original location and an insertion at the new location.
-
-
-Altered version 2
-```
-A,B
-3,12
-1,10
-2,11
-```
-
-Diff between original and altered version 2
-```diff
-A,B
-+3,12
-1,10
-2,11
--3,12
-```
-
-`write_vc()` avoids this by sorting the observations (rows) before writing them to the file. The names of the variables on which to sort and the sorting order are stored in the metadata. This information will be used each time `write_vc()` writes to the file.
-
[…]
