+++ vignettes/workflow.Rmd
@@ -1,0 +1,245 @@
+---
+title: "Potential workflows for working with dataframes under version control"
+author: "Thierry Onkelinx"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Potential workflows for working with dataframes under version control}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+  %\VignetteDepends{git2r}
+---
+
+```{r setup, include = FALSE}
+library(knitr)
+opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+set.seed(20120225)
+```
+
+## Introduction
+
+This vignette describes a potential workflow for storing dataframes. This time we use a `git2r::repository()` object as the root. This adds git functionality to `write_vc()` and `read_vc()`.
+
+The rationale behind this workflow is that we have read-only access to a database containing the raw data. The database is beyond our control. Observations in the database can be added, removed or updated without our knowledge. These changes cannot be traced in the database.
+
+The database defines a variable number of groups (e.g. species). We have defined a standard analysis which should run for each group. We want to repeat the analyses with some predefined frequency (e.g. once every year). In order to make the analyses reproducible, we want to store the relevant data in a git repository.
+
+## Setup
+
+We start by initializing a git repository. `git2rdata` assumes that is already done. Therefore we'll use the `git2r` functions to to so.
+
+```{r initialize}
+# initialize a bare git repo to be used as remote
+remote <- tempfile("git2rdata-workflow-remote")
+dir.create(remote)
+git2r::init(remote, bare = TRUE)
+
+# initialize a local git repo
+path <- tempfile("git2rdata-workflow")
+dir.create(path)
+init_repo <- git2r::clone(remote, path, progress = FALSE)
+git2r::config(init_repo, user.name = "me", user.email = "me@me.com")
+# add an initial commit with .gitignore file
+writeLines("C.*", file.path(path, ".gitignore"))
+git2r::add(init_repo, ".gitignore")
+git2r::commit(init_repo, message = "Initial commit")
+# push initial commit to remote
+git2r::push(init_repo, "origin", "refs/heads/master")
+```
+
+Next we need a function to mimic the data selection from the database.
+
+```{r dummy_data}
+generate_data <- function(x, n = rpois(1, 10)) {
+  if (missing(x)) {
+    coef <- runif(2)
+    sigma <- rchisq(1, 10) / 10
+    covariate <- runif(n)
+    old_covariate <- numeric(0)
+    old_response <- numeric(0)
+  } else {
+    coef <- attr(x, "coef")
+    sigma <- attr(x, "sigma")
+    x <- x[rbinom(nrow(x), size = 1, prob = 0.9) == 1, ]
+    update <- rbinom(nrow(x), size = 1, prob = 0.1) == 1
+    covariate <- c(x$covariate[update], runif(n))
+    old_covariate <- x$covariate[!update]
+    old_response <- x$response[!update] 
+  }
+  response <- coef[1] + covariate * coef[2] + 
+    rnorm(length(covariate), sd = sigma)
+  x <- data.frame(
+    covariate = c(old_covariate, covariate), 
+    response = c(old_response, response)
+  )
+  attr(x, "coef") <- coef
+  attr(x, "sigma") <- sigma
+  return(x)
+}
+```
+
+## Storing dataframes into a git repository
+
+### First commit
+
+Suppose that we have two groups at the first point in time. We read the data for these group from the database. We also store them in a list called `content` to be reused in the [next section](#automated-workflow-for-storing-dataframes).
+
[…]
