+++ _pkgdown.yml
@@ -11,6 +11,8 @@ navbar:
       href: articles/plain_text.html
     - text: Storing dataframes under version control
       href: articles/version_control.html
+    - text: Potential workflow
+      href: articles/workflow.html
     - text: Efficiency
       href: articles/efficiency.html
   - text: Functions
