### 2019-02-26 851c97e
remove auto_commit() because of hardly any extra functionality over git2r::commit()
