+++ man/prune_meta.Rd
@@ -51,7 +51,7 @@ git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
 status(repo)
 list_data(repo)
-auto_commit(package = "my_package", repo)
+commit(repo, "add iris data", session = TRUE)
 status(repo)
 rm_data(repo, path = ".")
 list_data(repo)
