### 2019-02-26 80dc7a4
update timings of efficiency vignette to new version
