+++ vignettes/efficiency.Rmd
@@ -192,7 +192,7 @@ The reduction in file size when storing in factors depends on the length of the 
 
 ```{r factor_label_length, echo = FALSE, fig.cap = "Effect of the label length on the efficiency of storing factor optimized, assuming 1000 observations", warning = FALSE}
 ratio <- function(label_length = 1:20, n_levels = 9, n_obs = 1000) {
-  meta_length <- 30 + (12 + label_length) * n_levels
+  meta_length <- 63 + (5 + label_length + floor(log10(n_levels))) * n_levels
   optimized <- n_obs * mean(ceiling(log10(seq_len(n_levels) + 1)))
   verbose <- n_obs * label_length
   ifelse(
@@ -342,7 +342,7 @@ if (system.file("efficiency", "git_size.rds", package = "git2rdata") == "") {
 
 Each version of the data has on purpose a random order of observations and variables. This is what would happen in a worst case scenario as it would generate the largest posibble diff. We also test `write.table()` with a stable ordering of the observations and variables. 
 
-The randomised `write.table()` yields the largest git repository, converging to about 6.5 times the size of a git repository based on the sorted `write.table()`. `saveRDS()` yields a 25% reduction in repostory size compared to the randomised `write.table()`, but still is almost 5 times larger than the sorted `write.table()`. Note that the gain of storing binary files in a git repository is much smaller than the gain in individual file size because the git repository will be compressed too. The optimized `write_vc()` starts at 81% converges toward 74%, the verbose version starts at 88% and converges towards 99%. There is a clear gain when using `write_vc()` with optimization in terms of storage size and the availability of metadata. The verbose option of `write_vc()` lack the gain in terms of storage size but still has the metadata advantange.
+The randomised `write.table()` yields the largest git repository, converging to about 6.5 times the size of a git repository based on the sorted `write.table()`. `saveRDS()` yields a 25% reduction in repostory size compared to the randomised `write.table()`, but still is almost 5 times larger than the sorted `write.table()`. Note that the gain of storing binary files in a git repository is much smaller than the gain in individual file size because the git repository will be compressed too. The optimized `write_vc()` starts at 83% converges toward 72%, the verbose version starts at 90% and converges towards 105%. There is a clear gain when using `write_vc()` with optimization in terms of storage size and the availability of metadata. The verbose option of `write_vc()` lack the gain in terms of storage size but still has the metadata advantange.
 
 ```{r plot_git_size, echo = FALSE, fig.cap = "Size of the git history using the different storage methods."}
 rs <- lapply(
@@ -411,7 +411,7 @@ if (system.file("efficiency", "file_timings.rds", package = "git2rdata") == "") 
 }
 ```
 
-`write_vc()` takes 75% to 100% more time than `write.table()` because it needs to prepare the metadata and sort the observations and variables. When overwriting existing data, the new data is checked against the existing metadata.
+`write_vc()` takes 67% to 86% more time than `write.table()` because it needs to prepare the metadata and sort the observations and variables. When overwriting existing data, the new data is checked against the existing metadata. `saveRDS()` requires only 44% of the time that `write.table()` needs.
 
 ```{r plot_file_timings, echo = FALSE, fig.cap = "Boxplot of the write timings for the different methods."}
 mb$expr <- reorder(mb$expr, mb$time, FUN = median)
@@ -445,7 +445,7 @@ if (system.file("efficiency", "read_timings.rds", package = "git2rdata") == "") 
 }
 ```
 
[…]
