### 2019-02-26 d6fc3f4
update README and add links to the vignettes. closes #12
