### 2019-02-26 c9c686b
add extra comments in the examples (#11)
