+++ R/recent_commit.R
@@ -10,18 +10,31 @@
 #' @export
 #' @family version_control
 #' @examples
+#' # initialise a git repo using git2r
 #' repo_path <- tempfile("git2rdata-repo")
 #' dir.create(repo_path)
 #' repo <- git2r::init(repo_path)
 #' git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
+#'
+#' # write and commit a first dataframe
 #' write_vc(iris[1:6, ], "iris", repo, sorting = "Sepal.Length", stage = TRUE)
 #' commit(repo, "important analysis", session = TRUE)
+#' list.files(repo_path)
 #' Sys.sleep(1.1) # required because git doesn't handle subsecond timings
+#'
+#' # write and commit a second dataframe
 #' write_vc(iris[7:12, ], "iris2", repo, sorting = "Sepal.Length", stage = TRUE)
 #' commit(repo, "important analysis", session = TRUE)
+#' list.files(repo_path)
 #' Sys.sleep(1.1) # required because git doesn't handle subsecond timings
+#'
+#' # write and commit a new version of the first dataframe
 #' write_vc(iris[7:12, ], "iris", repo, stage = TRUE)
+#' list.files(repo_path)
 #' commit(repo, "important analysis", session = TRUE)
+#'
+#' # find out in which commit a file was last changed
+#'
 #' # "iris.tsv" was last updated in the third commit
 #' recent_commit("iris.tsv", repo)
 #' # "iris.yml" was last updated in the first commit
@@ -30,9 +43,12 @@
 #' recent_commit("iris2.yml", repo)
 #' # the data object "iris" was last updated in the third commit
 #' recent_commit("iris", repo, data = TRUE)
+#'
+#' # remove a dataframe and commit it
 #' file.remove(file.path(repo_path, "iris.tsv"))
 #' prune_meta(repo, ".")
 #' commit(repo, message = "remove iris", all = TRUE, session = TRUE)
+#'
 #' # still points to the third commit as it is the latest commit in which the
 #' # data was present
 #' recent_commit("iris", repo, data = TRUE)
