### 2019-02-26 c764237
update NEWS file
