### 2019-02-28 0691dbc
Update README.md
Corrected some types
