### 2019-02-28 1263da8
increase subsecond sleep time to avoid bogus unit test error
