### 2019-02-28 d5debe9
add a warning that read_vc() can only read dataframe which are stored with write_vc(). As suggested in #15
