+++ R/read_vc.R
@@ -1,4 +1,6 @@
 #' Read a \code{data.frame} from a repository
+#'
+#' Note that the dataframe has to be written with `write_vc()` before it can be read with `read_vc()`.
 #' @inheritParams write_vc
 #' @return The \code{data.frame} with the file names and hashes as attributes
 #' @rdname read_vc
