### 2019-02-28 1772a68
update figures as suggested in #19
