### 2019-03-08 863bfb1
Readme: language enhancement and typos
