### 2019-03-08 00a3cd7
Readme: explicitly mention R
