### 2019-03-08 461f77a
list_data(), read_vc(), write_vc(): align titles with other functions *
For other functions, the explicit referral to a 'repository' is not made. In these functions, it is not needed either. The changes make the list of functions on the pkgdown website look more uniform.
