### 2019-03-08 6a77619
Vignettes: language enhancement and typos
