+++ vignettes/plain_text.Rmd
@@ -45,7 +45,7 @@ Storing the factors, POSIXct and Date as their index, makes them less user reada
 
 Another main goal of of `git2rdata` is to optimise the storage of the plain text files under version control. `write_vc()` and `read_vc()` has methods for interacting with [git](https://git-scm.com/) repositories using the `git2r` framework. Users who want to use git without `git2r` or use a different version control system (e.g. [Subversion](https://subversion.apache.org/), [Mercurial](https://www.mercurial-scm.org/)), still can use `git2rdata` to write the files to disk and uses their prefered workflow on version control. 
 
-Hence, `write_vc()` will always perform checks to look for changes which potentially lead to large diffs. More details on this in a dedicated vignette. Some problem will always yield a warning. Other problems will yield by default an error. The user can turn these errors into warnings by setting the `strict = FALSE` argument.
+Hence, `write_vc()` will always perform checks to look for changes which potentially lead to large diffs. More details on this in a dedicated vignette. Some problems will always yield a warning. Other problems will yield an error by default. The user can turn these errors into warnings by setting the `strict = FALSE` argument.
 
 As this vignette ignores the part on version control, we will always use `write_vc(strict = FALSE)` and hide the warnings to improve the readability.
 
@@ -130,7 +130,7 @@ all.equal(x, y2, check.attributes = FALSE)
 
 ## Missing values
 
-`write_vc()` has an `na` arguments which specifies the string which is used to indicate missing values. Because we avoid using quotes, this string must be different from any character value in the data. This include factor labels when the data is stored verbose. This is checked and will always return an error, even with `strict = FALSE`.
+`write_vc()` has an `na` argument which specifies the string which is used to indicate missing values. Because we avoid using quotes, this string must be different from any character value in the data. This include factor labels when the data is stored verbose. This is checked and will always return an error, even with `strict = FALSE`.
 
 ```{r echo = FALSE, results = "hide"}
 stopifnot("X" %in% x$x, "b" %in% x$y)
@@ -150,7 +150,7 @@ print_file("custom_na.tsv", path, 10)
 print_file("custom_na.yml", path, 4)
 ```
 
-The default string for missing values is `"NA"`. We recommend to keep this default, as long as the dataset permits it. A first good alternative is an empty string (`""`). If that won't work either, you'll have to use your imagination. Try to keep is short, clear and robust^[robust in the sense that you won't need to change it later].
+The default string for missing values is `"NA"`. We recommend to keep this default, as long as the dataset permits it. A first good alternative is an empty string (`""`). If that won't work either, you'll have to use your imagination. Try to keep it short, clear and robust^[robust in the sense that you won't need to change it later].
 
 ```{r empty_na}
 write_vc(x, "custom_na", path, strict = FALSE, na = "")
