### 2019-03-08 ed64dad
workflow.Rmd: bugfix commit(all=TRUE) in text
