### 2019-03-08 a60d7ff
efficiency.Rmd: reword levels as labels *
In this sentence, the word 'levels' was used twice, with confusing meaning. One occurrence is replaced as 'labels'.
