### 2019-03-12 f348a21
tweak vignettes as suggested by @stijnvanhoey and @florisvdh
