+++ tests/testthat/test_d_recent_commit.R
@@ -70,8 +70,13 @@ write_vc(
   sorting = "test_Date"
 )
 commit_7 <- commit(root, "second subsecond")
+write_vc(
+  test_data[15:16, ], file = "subsecond", root = root, stage = TRUE,
+  sorting = "test_Date"
+)
+commit_8 <- commit(root, "third subsecond")
 expect_warning(
   output <- recent_commit(file = "subsecond", root, data = TRUE),
   "Multiple commits within the same second"
 )
-expect_true(all(output$commit %in% c(commit_6$sha, commit_7$sha)))
+expect_true(all(output$commit %in% c(commit_6$sha, commit_7$sha, commit_8$sha)))
