### 2019-03-12 #263 @ThierryO
A new version (tagged [v0.0.3](https://github.com/inbo/git2rdata/tree/v0.0.3)) is ready for another round of review.
