### 2019-03-12 #263 @noamross
Thanks @ThierryO! @brodieG and @jtr13,  let us know if you are satisfied with these changes, including covering any areas you weren't able to in the last review.

@ThierryO, can you give a brief summary of changes? I notice there isn't a NEWS entry for v0.0.3, or is it effectively the same summary as 0.0.2?
