### 2019-04-11 #263 @brodieG
**Note**: one small change was needed to get package to build locally (see patch
where we change `git2r::add(..., force=TRUE)` because on my system I have a
global gitignore for dot files.)

A few "Package Guidelines" comments:

- Does the author of the `git2r` package welcome the use of `git2rdata` as this
  package name?
- Missing Ropensci Badge.

---

# Review Comments

## Overview

This is a useful and well written package.  Congratulations to Thierry for
crafting it.

The code is clean and clear.  For the most part, parameters are validated and
corner cases are accounted for.  Test coverage is complete (at least on a lines
of code basis).  Clearly Thierry is very conscientious and I am as confident as
I can be from static "eyeball" analysis alone that there are probably not too
many bugs lurking in the code outside of the potential ones I highlight in the issues
section next.  I also encountered no bugs in my manual testing.
This is far from a guarantee of course.

I see no obvious performance issues.

The documentation is extensive, particularly the vignettes.  I'm really
impressed by how seriously our initial round of feedback on the matter was
taken.  The function documentation is complete, particularly with the examples,
but it is a bit too terse (more details later).

## Issues That Should Be Addressed

These are important and my approval of the package is contingent on them being
addressed in some way:

* Indiscriminate remove of yml files (see `rm_data`/`prune_meta` comments, and
  File Storage comments)
* POSIXct time zone assumption (see `meta` comments)

Strongly recommended

* Related to the above, we recommend you adopt a more robust storage format (see
  File Storage comments)

Recommended:

* More prominent discussion of the performance/storage improvements from
  optimizing the data.  This is the key tangible selling point of the package.
  It should be a major part of the README.  You should also note the target data
  size (e.g  # obs < 1MM?).
* `do.call` and `sorting` corner cases for `meta` (see function section)
* Add the `write_vc` optimization suggested.

Every other comment in the review is provided for the author to decide whether
to implement or not.

## File Storage

In an earlier iteration I had suggested that files be stored in a subfolder for
convenience and future flexibility.  In reviewing the code and thinking of
additional validation required in `read_vc` and `prune_meta`, I am becoming
increasingly convinced that this achieves another important role: discouraging
non-vc "use" of the data to interfere with the vc function.  Two motivating
examples:

* Imagine the horror if a user opens a tsv file in Excel.
* What if the user has non vc yml files floating around?  Less likely but still
  possible, non-vc and yml and tsv files with the same name?

You could have something like:

iris.grx
+- UUID   (a UUID identifier you can use to check this is actually a vc dir)
+- meta   (this corresponds to iris.yml)
+- data   (this corresponds to iris.tsv)

`grx` is a made up scary sounding extension.  Obviously everything is still
stored in plain text so diffs, etc., continue to work fine.

Then `export_vc('iris.grx')` can pull out the `tsv` and `yml`.

This is a very cheap way to make your package more robust to file system
interactions you do not control.

## Typos

See attached patch: [git2rdata_typo_patch.txt](https://github.com/ropensci/software-review/files/3070453/git2rdata_typo_patch.txt)
## Vignettes

### Generally

Title case in headers, particularly the top level vignette title?  Not something
I particularly care about, but I hear some can be fastidious about it.

### Getting Started

Another frequently used csv read/write set of functions is
`data.table::fread/fwrite`.  Suffers the same factor loss infirmity as the other
functions reviewed.

In Re:

> POSIXct is written as an integer to the data. The class and the origin are
> stored in the metadata. Timestamps are always stored and returned as UTC.

POSIXct is numeric, not integer, consider:

```
> options(digits=22)
> as.numeric(as.POSIXct('2019-01-01 10:01:03.13413'))
[1] 1546354863.134130001068
```

I'm not sure this had any practical significance, although you may need to check
that sub-second precision is not lost, and if it is, at least document and warn
that this is happening when it happens.

Additionally, how comfortable are you with time zones and POSIXct?  I am not
all, and know just enough about them to be a little scared.  Have you confirmed
that timestamps with different timezones are correctly converted to UTC?  It
seems not (see comments in `meta` section).

### Suggested Workflow

#### Introduction

> The database defines a variable number of dataframe (e.g. species can be added
> or removed). We have defined a standard analysis which should run for each
> group. We want to repeat the analyses with some predefined frequency (e.g.
> once every year). In order to make the analyses reproducible, we want to store
> the relevant data in a git repository.

This paragraph might be confusing to some.  You should either generalize
further, or layout a clear schema of what the database contains.  Right now you
are at a middle level where you provide some detail, but not enough to be able
to infer the exact schema.  Does each table contain one species?  What is a
group in this context?  A species, or a sub-group in each table?

#### Storing Dataframes

See comments about `content` list below.

#### Automated Workflow

I find this section more ambitious than it needs to be.  It tries to
introduce several concepts at once, when perhaps the single most important one
is just the use of `store_data` with a single instance of the data.  `lapply`ing
over several instances of the data adds a layer of complexity that will make it
harder to understand the simple and more likely use case where there is just one
instance of data.  After all, aren't the different instances of the data
supposed to be tracked by the VCS system, rather than by a nested list in the
users workspaces?

Another drawback of the multi-instance example is that it interferes with the
previous instances.  The reader is left wondering what on earth this weird
nested list (`content`) we're creating is since we don't even use it until much
later.

I think you could introduce `store_data` ahead of the "Third Commit" and use
that for the third commit.  Text could be "After having done this two months in
a row, you are probably wondering if we could automate some of these steps...".
This also provides motivation for the "Automated workflow".  This may require
you to rework the Second commit to delete a data set so you have an example
to motivate the functionality of `store_data`.

Nit: `store_data` should at a minimum pluralize the `df` argument to make it
clear that it contains multiple data frames, or even better use `df_list`, or
alternatively, use `function(..., repo) {}`.

### Analysis Workflow

Great example!  One nit: maybe add newlines or otherwise try to better format
the meta data displayed ahead of the table?  It's a bit overwhelming with the
hashes and temp path names.  If at least each element is on its own line it will
be easier to see what's going on.

### Other Thoughts

#### Example Data Structure

Not a big deal at all, but it is very odd to me to see a database with tables of
identical structure that are only different in the dataset identifier.  This
would normally be stored as one table with a dataset column.

#### Associating Code With Data

Your footnote about "code changing" reminded me that it is useful to be able to
associate the data with the version of the code that ran it.  Perhaps the
`write_vc` function should warn when in `stage=TRUE` mode if there are other
unstaged changes?  This would improve the probability that scripts with analysis
residing in the same repository at the data are also committed.  It would be a
shame if analysis became irreproducible because the code that generated the
analysis was not committed along with the data.  This is not critical to address
at all.  Just a thought.

### Efficiency in Terms of Storage and Time

This is a great vignette.  Very informative.

You should promote the data storage statistics to a more prominent part of the
package, e.g. the README with something along the lines of:

> Data stored with `git2rdata` will generate git repositories x% smaller than if
> stored as RDSes, up to y% smaller than if written as is, etc...

The only additional bit of information that would be worthwhile to add is some
guidelines for how people should extrapolate file growth, e.g. made up:

> With a data frame of the same composition as `airbag` and in-memory size X,
> you can expect to use X * (file change % per commit) * (typical compression
> factor) * number of commits.  For `airbag` this translates to ~6MB over 100
> commits.  For reference the repository for <name of popular package with many
> commits> is ZZ MBs in size.

### Optimizing storage for ...

Good vignette.  See typos in patch.

## Generic Documentation Comments

### Descriptions

In general these are too terse.  Several do not describe what the function
actually does, e.g.:

* `read_vc`: Note that the dataframe has to be written with write_vc() before it
  can be read with read_vc().
* `meta`: write_vc applies this function automatically on your data.frame.

You are better off making the titles briefer and the description more
descriptive, e.g. for `meta` title becomes "Optimize Data for Storage", and
description goes into more detail of what happens to data frames and vectors,
and whether there are actually any user-facing use cases, or if this is just
exposed to the user for study or whatever.

### Nit Picks

Title Case.  While this is not strictly requested in Writing R Extensions, it is
universally done in the base packages.

Almost all sections of the documentation should be written in full sentences.
Make sure you have periods at the ends of your sentences, even for short
descriptions, parameters, return values.

See base functions such as `?mean` for stylistic cues.

### Common Parameters

@param x: "the 'data.frame" stray quote mark?
@param file: "the name of the file without file extension".  This could use more
explanation.  E.g. "the base file name shared by the tsv and yml files that
contain or will contain the version control ready data"

### Other

Examples should cleanup on exit, as per CRAN policies.  Currently some create
temporary files (e.g. `list_data`) that are not fully cleaned up (e.g. `repo`,
`root` dir).

Good use of "See Also" to link related functions.

For the re-exported functions, the display link would be best if it showed the
package name, e.g. `git2r::commit`, to make it completely obvious what is going
on.  It is slightly disorienting to land on the "commit" help page, only to be
told to go see `commit`.  Additionally, you should add a period after the
function name as "Description" is supposed to be at least a sentence.

## Functions

### Parameter Validation

It is great to see that the functions validate their parameters carefully.

One additional thing you could do, although this isn't common practice, is to
also check that the dispatching parameter actually is what it is supposed to be.
Just because a "character" method was called doesn't mean that the parameter is
character since in R you can call methods directly.

### list_data

In:

```
gsub(paste0("^", root, "/"), "", data_files)
```

There is the risk that `root` could contain regex characters with special
meaning.  You can use something like:

```
has.root <- substr(data_files, 1, nchar(root) + 1) == paste0(root, "/")
data_files[has.root] <-
  substr(data_files[has.root], nchar(root) + 1, nchar(data_files[has.root])
```

I have not tested this, but you get the idea.  Watch out for corner cases.  I
don't think `gsub(..., fixed=TRUE)` can work because you no longer can use the
start of string anchor.

### meta

#### Docs

The description should give some indication of what the function does, not just
that `write_vc` applies it.  Additionally, more detail on the `data.frame`
method is warranted (i.e. it applies `meta` on the columns).

#### Character Method

Good job of handling corner cases (e.g. strings containing quotes).

#### Factor Method

```
  if (isTRUE(optimize)) {
```

Can cause problems for users that might pass a "truthy" value expecting this to
work as is so often the case in R.  You should change the function to either
validate the parameters itself, or have more sophisticated handling of this
value.  I guess more generally the method should validate its input since it can
be called directly.

#### Logical Method

See Factor above.

#### POSIXct Method

I am not an expert on this, but assuming that the timezone is UTC causes
problems, e.g. on my system (OSX):

```
> x <- data.frame(as.POSIXct('1970-01-01'))
> write_vc(x, 'hello', strict=FALSE)
19c2fa6018df9fd9f6e6eb33ddd5f02d33383335
                             "hello.tsv"
9aca22d70c2aab43e6cece767909924d5c1d5a50
                             "hello.yml"
In meta.data.frame(x, optimize = optimize, na = na, sorting = sorting,  :
  no sorting applied
> read_vc('hello')
  as.POSIXct..1970.01.01..
1      1970-01-01 05:00:00
> as.POSIXct('1970-01-01')
[1] "1970-01-01 EST"
> Sys.timezone()
[1] "America/New_York"
>
```

This risks being a headache resolving.

#### Date Method

I don't think `tz="UTC"` does anything here?

```
    z <- format(x, format = "%Y-%m-%d", tz = "UTC")
```

#### data.frame Method

This function should check its parameters, sorting in particular.

That `..generic` is a reserved column name should be documented in `write_vc`
and also probably in `meta`.

In:

```
    x <- x[do.call(order, x[sorting]), , drop = FALSE] # nolint
```

You need `unname(x[sorting])` to avoid potential collisions between `order`
parameters and column names.

You need to handle the case where `sorting` is zero length (included in patches).

```
> meta(iris, sorting=character())
[1] Sepal.Length Sepal.Width  Petal.Length Petal.Width  Species
<0 rows> (or 0-length row.names)
```

Further down, in:

```
  if (has_name(dots, "old")) {
```

Probably not of great consequence, but it feels "wrong" to have an undocumented
internal optional parameter.  I imagine there aren't many if any situations that
will lead to a collision, but it would feel better to me to make it a formal
argument and document it, even if as `.old` and marked as "for internal use".

Finally, probably okay in this context because you overwrite `z`:

```
  z <- lapply(z, `attr<-`, "meta", NULL)
```

But you should be aware if you were not that calling setter functions indirectly
is not recommended as it is possible under some circumstances they will modify
the element by reference.

```
> x <- c(a=1)
> x
a
1
> y <- `names<-`(x, 'b')
> y
b
1
> x  # `x` changed by reference!
b
1
```

### read_vc

#### Docs

The description should describe what the function does, e.g. "Retrieves a
`data.frame` stored with `write_vc` from file or git repository storage."

It is not clear from the documentation that this is an S3 generic with methods.
Adding `@rdname read_vc` to the methods should mostly address this.  You may
also want to make it clear that the generic dispatches on the second argument,
which is quite uncommon (in fact I discovered it was possible when reviewing
this package).  You should also consider adding some commentary about how it
might be relatively straightforward to extend `git2rdata` to work with other
VCS systems by adding S3 methods for this generic, modeled on the
`git_repository` method.

#### File integrity

`read_vc` should take some measures to ensure that neither the data nor the meta
data are corrupted.  It seems reasonably easy that someone will either manually
modify one of the files intentionally, or unintentionally e.g. via git
checkout.

Some possible checks:

* Check that the meta data is in the expected structure, including:
    * all sub-elements of `..generic` are there and are what they should be.
    * all column definitions are there and are what they should be (i.e they all
      have a `class` element that is character, for factors the other elements
      are present, etc)
    * alternatively, check that the stored hash matches the data?  Not sure what
      you're using that hash for right now.  I think it is the hash of the meta
      data before the hash is added.
* Check that the data is consistent with the meta data.  Easy check would be to
  store a hash of the data in the meta data, which I don't think is done
  currently.  Look out for factors with indices undefined in the meta data.

### write_vc

That `..generic` is a reserved column name should be documented in `write_vc`
and also probably in `meta`.

```
write_vc(..., sorting=TRUE) # sort all columns?
```

#### Performance

`anyDuplicated` on a data frame is costly:

```
> library(ggplot2)
> treeprof::treeprof(fn <- write_vc(diamonds, 'diamonds', sorting=names(diamonds)))
Profiling
auto gc: running with gc() first
First run in 2.08 seconds
Looping to 5 seconds
Parsing Rprof
Done
Ticks: 1676; Iterations: 3; Time Per: 2.009 seconds; Time Total: 6.026 seconds; Time Ticks: 1.676

                                                              milliseconds
write_vc ---------------------------------------------------- : 2009 -   0
    write_vc.character -------------------------------------- : 2009 -   0
        meta ------------------------------------------------ : 1435 -   0
        |   meta.data.frame --------------------------------- : 1435 -   0
        |       anyDuplicated ------------------------------- : 1332 -   0
        |       |   anyDuplicated.data.frame ---------------- : 1332 -   0
        |       |       anyDuplicated ----------------------- : 1332 -   0
        |       |           do.call ------------------------- : 1332 -   0
        |       |               <Anonymous> ----------------- : 1332 -   0
        |       |                   mapply ------------------ : 1332 - 241
        |       as.data.frame ------------------------------- :   41 -   0
        |       |   as.data.frame.list ---------------------- :   41 -   0
        |       |       do.call ----------------------------- :   41 -   0
        |       |           <Anonymous> --------------------- :   41 -   0
        |       |               row.names<- ----------------- :   16 -   0
        |       |               |   row.names<-.data.frame -- :   16 -   0
        |       |               anyDuplicated --------------- :   14 -   0
        |       |               |   anyDuplicated.default --- :   14 -  14
        |       |               as.data.frame --------------- :   11 -   0
        |       |                   as.data.frame.integer --- :   11 -   2
        |       lapply -------------------------------------- :   31 -   0
        |       |   FUN ------------------------------------- :   31 -  11
        |       |       meta -------------------------------- :   14 -   0
        |       [ ------------------------------------------- :   31 -   0
        |           [.tbl_df -------------------------------- :   31 -   0
        |               do.call ----------------------------- :   26 -   0
        |                   <Anonymous> --------------------- :   26 -   0
        |                       do.call --------------------- :   25 -   0
        |                           order ------------------- :   25 -  25
        write.table ----------------------------------------- :  531 -   0
        |   .External2 -------------------------------------- :  526 - 526
        hashfile -------------------------------------------- :   41 -  41
```

If you'd like you can take advantage of the radix sort that the data.table
authors contributed to R, especially since you are already ordering the data:

```
  # license: GPL >= 2
duplicated2 <- function(dat) {# `dat` must be ordered!!!
[…]
