### 2019-04-17 #263 @jtr13
(I have additional suggestions listed below, but my approval is not contingent on their completion.)

#### Note

A reminder that I was not able to fully test all functionality claims.

### Followup from previous comments

#### README

I recommended updating the README to reflect the explanation of the problem that the software is designed to solve provided in the previous version of the vignette.  The README has been updated. I would suggest further editing the Rationale section so that it is framed in terms of problems and solutions.

Again, I see three main rationales:

1) version control of data

2) optimization of data storage

3) meta data storage to preserve data types, factor levels, etc.

Currently more attention is paid in the README to meta data (3) rather than version control (1) and optimization (2).

Finally, I recommend adding another section, perhaps called "Getting Started," to the README that provides a short summary of the contents of each vignette. As such, the theory/rationale for the package would be separate from the how-to.

#### Building vignettes

I mentioned that `build_vignettes = TRUE` wasn't working. The installation instructions were updated with clear instructions that include a workaround for installing the vignettes. (Changes are happening with the `remotes` package so I wouldn't be surprised if the method for installing vignettes changes again soon.  Mentioning as a heads-up to keep an eye on this, especially given the high quality of the vignettes -- important to make sure that they're seen!)

#### Examples

I noted that more comments would be helpful to guide the user through the examples. Clear examples with comments are now provided.

#### CONTRIBUTING file

I noted that there was no CONTRIBUTING file nor contributing guidelines in the README. A CONTRIBUTING file was added to the `.github` folder with a link in the README. (I think the root directory is a better place for CONTRIBUTING since it's more likely to be seen earlier; it won't pop up from `.github` until the contributor submits a pull request as far as I can tell.)

#### Rationale

I asked about the benefits of the optimization that the package provides. An excellent vignette, [Efficiency in terms of storage and time,](https://inbo.github.io/git2rdata/articles/efficiency.html) was added that addresses this issue in depth.

#### Package name

I suggested changing the package name. The author considered the issue and holds that the name is appropriate.

#### `rm_data()` and `prune_meta()`

I asked about the benefit of using `rm_data()`. The author responded: "The `rm_data()` and `prune_meta()` are useful in an automated workflow where changes in the raw data can lead to the removal of data sets. First apply `rm_data()` to remove all tsv files. Then write all the data using `write_vc()`. This is (re)create the tsv files which are still needed. Since all tsv were removed at the start, the ones we don't need any more are no longer present. The final step is to use `prune_meta()` to remove leftover yml files with no matching tsv file...."

Clearly `prune_meta()` serves a purpose in looking for matches. I am also convinced that `rm_data()` is useful; it's harder than I remembered to remove all files of a certain type from R. (Can't do `file.remove("*.tsv")` for example.) These functions have additional benefits such as defaulting to `recursive = TRUE`.

#### Sorting

I asked about an option to turn off sorting. The author responded: "Making the sorting optional is OK when not storing the files under version control. However, then it is up to the user to use the sorting when writing to a file path and doing the version control outside `git2rdata` diffs..." Currently a warning is given if no sorting parameter is supplied. I believe that previously an error was thrown; I think the warning is a better option here.

### New comments

#### `pull()` conflict

The `pull()` function reexported from `git2r` conflicts with `dplyr::pull()`. Not much you can do here, just flagging it.

#### Warnings

Should some of the warnings perhaps be messages instead? "Warning" seems too strong for example in this case:

``` r
library(git2rdata)
    write_vc(iris[1:6, ], file = "iris", sorting = "Sepal.Length")
#> 11d423327dc1f67e16183dbd9871ad931d3b9689
#>                               "iris.tsv"
#> fb962b490dbe4038512e52b72b8182e11b49a60d
#>                               "iris.yml"
    write_vc(
        iris[1:6, -2], "iris", sorting = "Sepal.Length", strict = FALSE
    )
#> Warning in write_vc.character(iris[1:6, -2], "iris", sorting = "Sepal.Length", : new data has a different number of variables
#> deleted variables: Sepal.Width
#> 1a649561f36566acf21c2f167e69abc54907badb
#>                               "iris.tsv"
#> af0219b9d8ebdea4d30c833b14d3fadc22a2f750
#>                               "iris.yml"
```

<sup>Created on 2019-04-17 by the [reprex package](https://reprex.tidyverse.org) (v0.2.1.9000)</sup>

#### Error messages

It appears to be common practice to include warnings with errors though it seems to me it would be better not to get the warnings, so that the error isn't lost in the shuffle. For example:

```
> write_vc(iris[1,], "iris")
1d7d77ee4b27932996ebcfcba2c25ec9c5aa7630
                              "iris.tsv"
38fdea6690db69a1da66fa5f06f47b37cf3cfccc
                              "iris.yml"
Warning message:
In meta.data.frame(x, optimize = optimize, na = na, sorting = sorting) :
  no sorting applied

> write_vc(iris[,2:4], "iris" )
Error: new data has a different number of variables
deleted variables: Sepal.Length, Species
In addition: Warning message:
In meta.data.frame(x, optimize = optimize, na = na, sorting = sorting,  :
  no sorting applied
```

Also, it would be helpful to get a clear message that on error the data file was not changed.

#### Documentation

`read_vc()` has the same examples as `write_vc()`, which makes sense since they work together though you could cut the `write_vc()` examples that aren't followed by `read_vc()`

#### Vignettes

"Getting started"

Small comment:

```
library(git2rdata)
write_vc(x = x, file = "first_test", root = path, strict = FALSE)
```

produces the "no sorting applied" warning described above. It would be helpful if a link were provided here to the optimization vignette where this is clearly explained. (I tend to think something went wrong if I get a warning running vignette code without a heads-up.)

Overall, the vignettes are incredibly informative and add great value to the package. As noted above, it would be helpful to provide short summaries of the vignettes in the README, with a suggested order, such as:

1. "Getting Started": ...

2. "Optimizing storage for version control" ...

3. "Suggested workflow for storing a variable set of dataframes under version control" ...

4. "Efficiency in terms of storage and time" ... performance benchmarks...
