### 2019-04-18 #263 @ThierryO
Thanks to both @brodieG and @jtr13 for your detailed reviews. Your comments allow me to improve the quality of the package. I'll look into them and notify you when a next version is available. I'll giving an oral presentation on `git2rdata` at useR!2019. It would be nice to have the package on rOpenSci by then.

Below is a response on a few issues raised by @brodieG. Can I get some feedback on this as some these potentially require a major overhaul of the package.

## General issues

- The idea started as PR ropensci/git2r#366 in `git2r`. We [mentioned](https://github.com/ropensci/git2r/pull/366#issuecomment-419113376) the name `git2rdata` and the author of `git2r` gave a thumbs up.
- Peer review badge. I'll add this. I though we should add it after the package was accepted.

## File format

Using a custom file format has indeed the benefit that the user is less likely to tamper with the (meta)data. However, a [comment](https://github.com/ThierryO/git2r/pull/1#discussion_r206073017) on an earlier version explicit asks to store the data in an open standard format.

IMHO, the benefits using a clear open standard outweighs the fact to make it tamper proof. Image the case in which the data was send to someone without knowledge of `git2rdata`, or somebody who wants to use the data with other software. This is doable with the current format. Using unknown 'scary sounding' extensions will make that much harder, even when the data is stored in plain text format.

`git2rdata` does not aim to avoid any tampering with the data. It rather make it easier to document the history using a VCS.

Currently, I haven't tested how `read_vc()` reacts when the data and metadata no longer match. Some checks are easy to implement in `read_vc()` e.g. changes in the metadata, new variables added to the data, missing variables in the data, new order of the variables in the data, order of the observations, ...

I'm not sure if we should prevent user to make changes to the (information content of the) data itself using other software. Think of using the Github website to correct typo's in the data, to remove observations from the data or to even add a few observations. If these changes corrupt the data, the user can revert them use the VCS or the backup.

## metadata pruning

The `rm_data()` and `prune_meta()` combo are only relevant in a workflow using a VSC in which a variable number of object are written. Using these functions in combination with a VSC is semi-safe as the user can revert the changes.

I could make these functions safer by adding something like an `git2rdata.index`. Such file would record all files written by `git2rdata`. `rm_data()` and `prune_meta()` can use this information to limit the removal of file to those created by `git2rdata`.

## Time zone issue

IMHO, the issue is rather that `print.data.frame` doesn't show the time zone.

    library(git2rdata)
    x <- data.frame(TS = as.POSIXct('1970-01-01'))
    fn <- write_vc(x, 'hello')

    ## sorting, : no sorting applied

    z <- read_vc('hello')
    z

    ## 1 1969-12-31 23:00:00

    z$TS

    ## [1] "1969-12-31 23:00:00 UTC"

    Sys.timezone()
