### 2019-04-18 #263 @brodieG
## Badge

For the badge I'm going off of [the guide](https://ropensci.github.io/dev_guide/building.html):

> Badges for continuous integration and test coverage, the badge for rOpenSci peer-review once it has started (see below)

## File Format

It's not critical that you have a custom "format" for git2rdata.  I will note though that the "format" I propose still stores the data in open format in a completely accessible way.  Diffs will show it, CSV readers will be able to read it, etc.  I'm pretty sure that the concern would only really exist in the case where the above isn't true, where you need specialized tools to access the data.

Fundamentally the git2rdata object is a new non-standard format as it is the combination of the data and the meta data.  By putting the files together in a folder you are making that format explicit and giving the signal to users that if you mess with the data, you may well break the git2rdata object.  You can even have a README.txt in the folder explaining this instead of using a scary extension.  This is probably better anyway, and will allow recipients access to the data with the appropriate warning.

Finally, if you still prefer to have the yml and tsv files floating around, I won't stop you so long as you take measures to ensure that the file pruning and deletion only affects files that were created by `git2rdata`.  Do note that there is an inherent fragility at having an additional meta data file floating around.  What happens if someone sends you a new data/meta-data set?  If you have a folder then you are all set.  An alternative would be to add a UUID in the yml so that you can be sure (or at least very certain) that the yml actually belongs to git2rdata.  Additionally, it will be particularly important that you check for data / meta-data corruption, although I would personally put those checks in under any implementation.

To recap, based on your feedback and further thought, this is what I would have the folder look like for e.g. `iris`:

```
iris (or maybe iris.git2rdata)
+ README.txt
+ git2rdata.version    # text file with git2rdata version, see below
+ iris.tsv
+ iris.yml                     # contains hash of data file
```

The `git2rdata.version` file will allow you to make changes in the future to the directory format as you can check it on read, and if necessary, convert to the new format.
