### 2019-04-18 da6ca5f
fix typos as suggested by @brodieG
