+++ R/relabel.R
@@ -1,6 +1,6 @@
 #' Relabel factor levels
 #'
-#' Imaging the situation where we have a dataframe with a factor variable and we
+#' Imagine the situation where we have a dataframe with a factor variable and we
 #' have stored it with `write_vc(optimize = TRUE)`. The raw data file contains
 #' the factor indices and the metadata contains the link between the factor
 #' index and the corresponding label.
