+++ man/meta.Rd
@@ -29,7 +29,8 @@ meta(x, ...)
 
 \item{na}{the string to use for missing values in the data.}
 
-\item{optimize}{recode the data to get smaller text files. Defaults to TRUE}
+\item{optimize}{recode and re-order the data to get smaller text files.
+Defaults to TRUE}
 
 \item{index}{an optional named vector with existing factor indices. The names must match the existing factor levels. Unmatched levels from \code{x} will get new indices.}
 }
