### 2019-04-23 inbo/git2rdata#12 @ThierryO
- [X] add ROpenSci review badge
- [ ] move lifecycle badge from "experimental" to "maturing"
- [ ] move status badge from "WIP" to "Active"
- [ ] add short summary of the vignettes to the README
- [ ] add some rough guidelines about maximal size of the data
