### 2019-04-23 inbo/git2rdata#27 @ThierryO
Update meta()
- [x] replace `if(isTRUE(x))` with `assert_that(is.flag(x), noNA(x)); if(x)`
- [x] add link to optimisation vignette in case of "no sorting" warning
- [x] improve error messages of `meta()`
- [x] use alternative for `anyDuplicated()` as suggested in  https://github.com/ropensci/software-review/issues/263#issuecomment-482317349
- [x] drop the `tz` argument in `meta.Date()`
- [x] mark `..generic` as a reserved keyword
