### 2019-04-26 4d1639a
delete superfluous word 'of'
or is something lacking after the word of?
