### 2019-04-26 6fd270c
use is.flag() and noNA() instead of isTRUE()
