### 2019-04-26 8edb2db
mention the efficiency vignette with the sorting param
