### 2019-04-26 7433eb8
improve error messages in write_vc()
