### 2019-04-26 48e12f7
use more efficient alternative of anyDuplicated() as suggested by @brodieG
