### 2019-04-29 5fc2f80
write_vc() handles factors with empty labels ("") correctly #24
