### 2019-04-30 b5472df
relabel() handles empty labels ("") correctly. Fixes #24
