### 2019-04-30 ac7dd7c
update documentation of relabel()
