### 2019-05-02 a9927d8
read_vc() ignore the comment character
