### 2019-05-02 61f3178
read_vc() validates the metadata
