### 2019-05-02 facf49c
clean files written by unit tests
