+++ tests/testthat/test_e_empty_label.R
@@ -97,3 +97,5 @@ test_that("relabel handles empty labels", {
   relabel(file = file, root = root, change = change)
   expect_is(mydfr <- read_vc(file = file, root = root), "data.frame")
 })
+
+file.remove(list.files(root, recursive = TRUE, full.names = TRUE))
