+++ tests/testthat/test_a_basics.R
@@ -270,6 +270,7 @@ test_that("user specified na strings work", {
     grep("junk", readLines(file.path(root, fn[1]))),
     2:4
   )
+  file.remove(list.files(root, recursive = TRUE, full.names = TRUE))
 })
 
 test_that("write_vc() allows changes in factor levels", {
@@ -288,6 +289,7 @@ test_that("write_vc() allows changes in factor levels", {
     write_vc(x, "factor_levels", root),
     "New factor labels for 'test_factor'"
   )
+  file.remove(list.files(root, recursive = TRUE, full.names = TRUE))
 })
 
 test_that("meta attributes are printed as yaml", {
