+++ tests/testthat/test_e_validate_metadata.R
@@ -41,3 +41,4 @@ test_that("read_vc() handles changes in rawdata", {
   )
 })
 
+file.remove(list.files(root, recursive = TRUE, full.names = TRUE))
