+++ man/read_vc.Rd
@@ -8,7 +8,7 @@ read_vc(file, root = ".")
 }
 \arguments{
 \item{file}{the name of the file without file extension. Can include a
-relative path. It is relative to the \code{root}.}
+relative path. \code{file} is a path relative to the \code{root}.}
 
 \item{root}{The root of a project. Can be a file path or a \code{git-repository}.
 Defaults to the current working directory (".").}
