### 2019-05-03 296583c
add a data hash to the metadata
