### 2019-05-03 42f2280
store the git2rdata version number in the meta data. #25
