+++ R/relabel.R
@@ -96,6 +96,7 @@ Use write_vc() instead.")
     )
   }
   old_meta <- meta_data
+  old_meta[["..generic"]][["git2rdata"]] <- NULL
   old_meta[["..generic"]][["hash"]] <- NULL
   old_meta[["..generic"]][["data_hash"]] <- NULL
   meta_data[["..generic"]][["hash"]] <- hash(as.yaml(old_meta))
