+++ NAMESPACE
@@ -64,6 +64,7 @@ importFrom(git2r,status)
 importFrom(git2r,workdir)
 importFrom(methods,setOldClass)
 importFrom(stats,setNames)
+importFrom(utils,packageVersion)
 importFrom(utils,read.table)
 importFrom(utils,write.table)
 importFrom(yaml,as.yaml)
