### 2019-05-03 dae1ed2
update grammar in NEWS file
