### 2019-05-04 7261f3b
the metadata contains the git2rdata version number
