+++ tests/testthat/test_e_upgrade.R
@@ -1,0 +1,128 @@
+context("upgrade to new version")
+root <- tempfile("git2rdata-upgrade")
+dir.create(root)
+test_that("read_vc() checks version", {
+  file <- basename(tempfile(tmpdir = root))
+  junk <- write_vc(test_data, file = file, root = root, sorting = "test_Date")
+  correct_yaml <- yaml::read_yaml(file.path(root, junk[2]))
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["git2rdata"]] <- "0.0.3"
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    read_vc(file = file, root = root),
+    "Data stored using an older version of `git2rdata`."
+  )
+
+  junk_yaml[["..generic"]][["git2rdata"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    read_vc(file = file, root = root),
+    "Data stored using an older version of `git2rdata`."
+  )
+})
+
+test_that("relabel() checks version", {
+  file <- basename(tempfile(tmpdir = root))
+  junk <- write_vc(test_data, file = file, root = root, sorting = "test_Date")
+  new_labels <- list(test_factor = list(a = "xyz"))
+  correct_yaml <- yaml::read_yaml(file.path(root, junk[2]))
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["git2rdata"]] <- "0.0.3"
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    relabel(file = file, root = root, change = new_labels),
+    "Data stored using an older version of `git2rdata`."
+  )
+
+  junk_yaml[["..generic"]][["git2rdata"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    relabel(file = file, root = root, change = new_labels),
+    "Data stored using an older version of `git2rdata`."
+  )
+})
+
+test_that("upgrade_data() validates metadata", {
+  file <- basename(tempfile(tmpdir = root))
+  junk <- write_vc(test_data, file = file, root = root, sorting = "test_Date")
+  expect_error(
+    upgrade_data(file = file, root = pi),
+    "a 'root' of class numeric is not supported"
+  )
+
+  correct_yaml <- yaml::read_yaml(file.path(root, junk[2]))
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["git2rdata"]] <- NULL
+  junk_yaml[["test_Date"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    upgrade_data(file = file, root = root),
+    "Corrupt metadata, mismatching hash."
+  )
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["git2rdata"]] <- NULL
+  junk_yaml[["..generic"]][["hash"]] <- "zzz"
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    upgrade_data(file = file, root = root),
+    "Corrupt metadata, mismatching hash."
+  )
+  junk_yaml[["..generic"]][["hash"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_error(
+    upgrade_data(file = file, root = root),
+    "Corrupt metadata, no hash found."
+  )
+})
+
+file.remove(list.files(root, recursive = TRUE, full.names = TRUE))
+
+test_that("upgrade_data() works from 0.0.3 to 0.0.4", {
+  file <- basename(tempfile(tmpdir = root))
[…]
