+++ tests/testthat/test_e_non_ascii.R
@@ -1,4 +1,4 @@
-context("check writing special characters")
+context("check writing non ASCII characters")
 root <- tempfile("git2rdata-empty-label")
 dir.create(root)
 characters <- data.frame(a = c("€$£ @&#§µ^ ()[]{}|²³<>/\\*+- ,;:.?!~",
