### 2019-05-05 10cf118
handle root containing regex. Fixes #30.
