### 2019-05-05 07a963e
make unit test faster to make sure that at least two commits are within the same second
