### 2019-05-06 4f7c22d
add is_git2rmeta()
