+++ tests/testthat/test_d_is_git2rmeta.R
@@ -1,0 +1,68 @@
+context("validate metadata")
+root <- tempfile("git2rdata-is_git2rmeta(.*)")
+dir.create(root)
+test_that("is_git2rmeta checks root", {
+  expect_error(is_git2rmeta(file = "junk", root = 1),
+               "a 'root' of class numeric is not supported")
+})
+
+test_that("is_git2rmeta checks metadata", {
+  expect_false(is_git2rmeta(file = "junk", root = root))
+  expect_error(is_git2rmeta(file = "junk", root = root, validate = TRUE),
+               "Metadata file missing.")
+
+  file <- basename(tempfile(tmpdir = root))
+  junk <- write_vc(test_data, file = file, root = root, sorting = "test_Date")
+  correct_yaml <- yaml::read_yaml(file.path(root, junk[2]))
+
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_false(is_git2rmeta(file = file, root = root))
+  expect_error(is_git2rmeta(file = file, root = root, validate = TRUE),
+               "No '..generic' element.")
+
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["hash"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_false(is_git2rmeta(file = file, root = root))
+  expect_error(is_git2rmeta(file = file, root = root, validate = TRUE),
+               "Corrupt metadata, no hash found.")
+
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["git2rdata"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_false(is_git2rmeta(file = file, root = root))
+  expect_error(is_git2rmeta(file = file, root = root, validate = TRUE),
+               "Data stored using an older version of `git2rdata`.")
+
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["git2rdata"]] <- "0.0.3"
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_false(is_git2rmeta(file = file, root = root))
+  expect_error(is_git2rmeta(file = file, root = root, validate = TRUE),
+               "Data stored using an older version of `git2rdata`.")
+
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["data_hash"]] <- NULL
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_false(is_git2rmeta(file = file, root = root))
+  expect_error(is_git2rmeta(file = file, root = root, validate = TRUE),
+               "Corrupt metadata, no data hash found.")
+
+  junk_yaml <- correct_yaml
+  junk_yaml[["..generic"]][["hash"]] <- "zzz"
+  yaml::write_yaml(junk_yaml, file.path(root, junk[2]))
+  expect_false(is_git2rmeta(file = file, root = root))
+  expect_error(is_git2rmeta(file = file, root = root, validate = TRUE),
+               "Corrupt metadata, mismatching hash.")
+})
+
+root <- git2r::init(root)
+git2r::config(root, user.name = "Alice", user.email = "alice@example.org")
+
+test_that("is_git2rmeta handle git repositories", {
+  file <- basename(tempfile(tmpdir = git2r::workdir(root)))
+  junk <- write_vc(test_data, file = file, root = root, sorting = "test_Date")
+  expect_true(is_git2rmeta(file = file, root = root))
+})
