### 2019-05-06 75a40c5
list_data() limits the data files to those with valid metadata. #26
