+++ R/list_data.R
@@ -1,4 +1,7 @@
-#' List available data files
+#' List available git2rdata files containing data
+#'
+#' The selection lists all `.tsv` files with a matching **valid** metadata file.
+#' **invalid** metadata files result in a warning.
 #' @param root the `root` of the repository. Either a path or a `git-repository`
 #' @param path relative `path` from the `root`. Defaults to the `root`
 #' @inheritParams base::list.files
@@ -22,12 +25,21 @@ list_data.character <- function(root = ".", path = ".", recursive = TRUE) {
   root <- normalizePath(root, winslash = "/", mustWork = TRUE)
   path <- normalizePath(file.path(root, path), winslash = "/", mustWork = TRUE)
 
-  data_files <- list.files(path, pattern = "\\.tsv$", recursive = TRUE,
+  data_files <- list.files(path, pattern = "\\.tsv$", recursive = recursive,
                            full.names = TRUE)
-  meta_files <- list.files(path, pattern = "\\.yml$", recursive = TRUE,
+  meta_files <- list.files(path, pattern = "\\.yml$", recursive = recursive,
                            full.names = TRUE)
   data_files <- gsub("\\.tsv$", "", data_files)
   meta_files <- gsub("\\.yml$", "", meta_files)
+  meta_files <- meta_files[meta_files %in% data_files]
+  meta_files_base <- remove_root(file = meta_files, root = root)
+  check <- vapply(X = meta_files_base, FUN = is_git2rmeta,
+                  FUN.VALUE = NA, root = root, validate = FALSE)
+  if (any(!check)) {
+    warning("Invalid metadata files found. See ?is_git2rmeta():\n",
+            paste(meta_files_base[!check], collapse = "\n"))
+  }
+  meta_files <- meta_files[check]
   data_files <- data_files[data_files %in% meta_files]
   remove_root(file = data_files, root = root)
 }
