### 2019-05-06 bb06c66
prune_meta() removes only valid metadata #26
