### 2019-05-06 60b0b1e
add note about ..generic to write_vc()
