### 2019-05-07 63e6e65
add is_git2rdata()
