### 2019-05-07 f4b5fa8
clean up files written by examples
