+++ man-roxygen/example-prune.R
@@ -53,3 +53,12 @@
 #' # check the removal
 #' list_data(repo)
 #' status(repo)
+#'
+#' # clean up
+#' junk <- file.remove(
+#'   list.files(root, full.names = TRUE, recursive = TRUE), root)
+#' junk <- file.remove(
+#'   rev(list.files(repo_path, full.names = TRUE, recursive = TRUE,
+#'                  include.dirs = TRUE, all.files = TRUE)),
+#'   repo_path)
+
