+++ man-roxygen/example-io.R
@@ -67,3 +67,11 @@
 #'   iris[1:6, ], "iris2", repo, sorting = "Sepal.Width", optimize = FALSE
 #' )
 #' status(repo)
+#'
+#' # clean up
+#' junk <- file.remove(
+#'   list.files(root, full.names = TRUE, recursive = TRUE), root)
+#' junk <- file.remove(
+#'   rev(list.files(repo_path, full.names = TRUE, recursive = TRUE,
+#'                  include.dirs = TRUE, all.files = TRUE)),
+#'   repo_path)
