+++ R/recent_commit.R
@@ -52,6 +52,12 @@
 #' # still points to the third commit as it is the latest commit in which the
 #' # data was present
 #' recent_commit("iris", repo, data = TRUE)
+#'
+#' #' clean up
+#' junk <- file.remove(
+#'   rev(list.files(repo_path, full.names = TRUE, recursive = TRUE,
+#'                  include.dirs = TRUE, all.files = TRUE)),
+#'   repo_path)
 recent_commit <- function(file, root, data = FALSE){
   UseMethod("recent_commit", root)
 }
