+++ man/prune_meta.Rd
@@ -93,6 +93,14 @@ prune_meta(repo, path = ".")
 # check the removal
 list_data(repo)
 status(repo)
+
+# clean up
+junk <- file.remove(
+  list.files(root, full.names = TRUE, recursive = TRUE), root)
+junk <- file.remove(
+  rev(list.files(repo_path, full.names = TRUE, recursive = TRUE,
+                 include.dirs = TRUE, all.files = TRUE)),
+  repo_path)
 }
 \seealso{
 Other storage: \code{\link{list_data}},
