### 2019-05-07 88272bc
make clear in the documentation of re-exported function that the documentation point to the originating package
