+++ R/reexport.R
@@ -1,33 +1,33 @@
 #' Reexported function from git2r
 #'
-#' See \code{\link[git2r]{repository}}
+#' See \code{\link[git2r]{repository}} in `git2r`
 #' @name repository
 #' @importFrom git2r repository
 #' @family version_control
 #' @export
 NULL
 
-#' Reexported function from git2r
+#' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{status}}
+#' See \code{\link[git2r]{status}} in `git2r`
 #' @name status
 #' @importFrom git2r status
 #' @family version_control
 #' @export
 NULL
 
-#' Reexported function from git2r
+#' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{commit}}
+#' See \code{\link[git2r]{commit}} in `git2r`
 #' @name commit
 #' @importFrom git2r commit
 #' @family version_control
 #' @export
 NULL
 
-#' Reexported function from git2r
+#' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{pull}}
+#' See \code{\link[git2r]{pull}} in `git2r`
 #' @name pull
 #' @importFrom git2r pull
 #' @family version_control
