+++ man/repository.Rd
@@ -4,7 +4,7 @@
 \alias{repository}
 \title{Reexported function from git2r}
 \description{
-See \code{\link[git2r]{repository}}
+See \code{\link[git2r]{repository}} in \code{git2r}
 }
 \seealso{
 Other version_control: \code{\link{commit}},
