### 2019-05-09 bafd96b
improve documentation
