### 2019-05-09 e290d26
upgrade_data() gains a path argument
