### 2019-05-09 5dd62e2
more improvements on the documentation
