+++ man-roxygen/example-io.R
@@ -7,23 +7,24 @@
 #'
 #' # write a dataframe to the directory
 #' write_vc(iris[1:6, ], file = "iris", root = root, sorting = "Sepal.Length")
-#' # check that a data file (.tsv) and a metadata file (.yml) are created
+#' # check that a data file (.tsv) and a metadata file (.yml) exist.
 #' list.files(root, recursive = TRUE)
-#' # read the data from the directory
+#' # read the git2rdata object from the directory
 #' read_vc("iris", root)
 #'
-#' # store a new version
+#' # store a new version with different observations but the same metadata
 #' write_vc(iris[1:5, ], "iris", root)
 #' list.files(root, recursive = TRUE)
-#' # store a new version in case the metamata must change
+#' # Removing a column requires version requires new metadata.
+#' # Add strict = FALSE to override the existing metadata.
 #' write_vc(
 #'   iris[1:6, -2], "iris", root, sorting = "Sepal.Length", strict = FALSE
 #' )
 #' list.files(root, recursive = TRUE)
-#' # storing the first version again required another update of the metadata
+#' # storing the orignal version again requires another update of the metadata
 #' write_vc(iris[1:6, ], "iris", root, sorting = "Sepal.Width", strict = FALSE)
 #' list.files(root, recursive = TRUE)
-#' # storing the data in verbose format leads to larger files
+#' # optimize = FALSE stores the data more verbose. This requires larger files.
 #' write_vc(
 #'   iris[1:6, ], "iris2", root, sorting = "Sepal.Width", optimize = FALSE
 #' )
@@ -31,7 +32,7 @@
 #'
 #'
 #'
-#' ## on git repo
+#' ## on git repo using a git2r::git-repository
 #'
 #' # initialise a git repo using the git2r package
 #' repo_path <- tempfile("git2rdata-repo-")
@@ -39,30 +40,18 @@
 #' repo <- git2r::init(repo_path)
 #' git2r::config(repo, user.name = "Alice", user.email = "alice@example.org")
 #'
-#' # store a dataframe in git repo
+#' # store a dataframe in git repo.
 #' write_vc(iris[1:6, ], file = "iris", root = repo, sorting = "Sepal.Length")
+#' # This git2rdata object is not staged by default.
 #' status(repo)
 #' # read a dataframe from a git repo
 #' read_vc("iris", repo)
 #'
-#' # store a new version in the git repo
+#' # store a new version in the git repo and stage it in one go
 #' write_vc(iris[1:5, ], "iris", repo, stage = TRUE)
 #' status(repo)
 #'
-#' # store a version with altered metadata
-#' write_vc(
-#'   iris[1:6, -2], "iris", repo, sorting = "Sepal.Length", strict = FALSE
-#' )
-#' status(repo)
-#'
-#' # store the original version again
-#' write_vc(
-#'   iris[1:6, ], "iris", repo, sorting = "Sepal.Width", strict = FALSE,
-#'   stage = TRUE
-#' )
-#' status(repo)
-#'
-#' # store a verbose version in separate files
+#' # store a verbose version in a different gir2data object
 #' write_vc(
 #'   iris[1:6, ], "iris2", repo, sorting = "Sepal.Width", optimize = FALSE
 #' )
