### 2019-05-13 d37d73b
bugfix in relabel() example
