### 2019-05-14 60cf530
upgrade workflow vignette
