+++ vignettes/workflow.Rmd
@@ -20,15 +20,20 @@ set.seed(20120225)
 
 ## Introduction
 
-This vignette describes a suggested workflow for storing dataframes. This time we use a `git2r::repository()` object as the root. This adds git functionality to `write_vc()` and `read_vc()`, provided by the [`git2r`](https://cran.r-project.org/package=git2r) package. This allows to pull, stage, commit and push from within R.
+This vignette describes a suggested workflow for storing a snapshot of dataframes as git2rdata objects under version control. The workflow comes in two flavours: 
 
-The rationale behind this workflow is that we have read-only access to a database containing the raw data. The database is beyond our control. Observations in the database can be added, removed or updated without our knowledge. These changes cannot be traced in the database.
+  1. A single repository holding both the data and the analysis code. The single repository set-up is simple. A single reference (e.g. commit) points to both the data and the code. 
+  1. One repository holding the data and a second repository holding the code. The data and the code have an independent history under a two repository set-up. Documenting the analysis requires one reference to each repository. Such a set-up is useful for repeating the same analysis (stable code) on updated data.
 
-The database defines a variable number of dataframes (e.g. species can be added or removed). We have defined a standard analysis which should run for each group. We want to repeat the analyses with some predefined frequency (e.g. once every year). In order to make the analyses reproducible, we want to store the relevant data in a git repository.
+In this vignette we use a `git2r::repository()` object as the root. This adds git functionality to `write_vc()` and `read_vc()`, provided by the [`git2r`](https://cran.r-project.org/package=git2r) package. This allows to pull, stage, commit and push from within R.
+
+Each commit in the data git repository describes a complete snapshot of the data at the time of the commit. The difference between two commits can consist of changes in existing git2rdata object (updated observations, new observations, deleted observations or updated metadata). Besides updating the existing git2rdata objects, we can also add new git2rdata objects or remove existing ones. Such higher level addition and deletions need to be tracked as well.
+
+We illustrate the workflow with a mock analysis on the `datasets::beaver1` and `datasets::beaver2` datasets.
 
 ## Setup
 
-We start by initializing a git repository. `git2rdata` assumes that is already done. Therefore we'll use the `git2r` functions to to so.
+We start by initializing a git repository. `git2rdata` assumes that is already done. Therefore we'll use the `git2r` functions to do so. We start by creating a local bare repository. In practice we will use a remote on an external server (GitHub, Gitlab, Bitbucket, ...). The example below creates a local git repository with an upstream git repository. Any other workflow to create a similar structure is fine.
 
 ```{r initialize}
 # initialize a bare git repo to be used as remote
@@ -42,154 +47,168 @@ dir.create(path)
 init_repo <- git2r::clone(remote, path, progress = FALSE)
 git2r::config(init_repo, user.name = "me", user.email = "me@me.com")
 # add an initial commit with .gitignore file
-writeLines("C.*", file.path(path, ".gitignore"))
-git2r::add(init_repo, ".gitignore", force=TRUE)
+writeLines("*extra*", file.path(path, ".gitignore"))
+git2r::add(init_repo, ".gitignore", force = TRUE)
 git2r::commit(init_repo, message = "Initial commit")
 # push initial commit to remote
 git2r::push(init_repo, "origin", "refs/heads/master")
+rm(init_repo)
 ```
 
-Next we need a function to mimic the data selection from the database.
+## Structuring git2rdata objects within a project
 
[…]
