+++ NEWS.md
@@ -1,4 +1,4 @@
-git2rdata 0.0.4 (2019-05-06)
+git2rdata 0.0.4 (2019-05-14)
 ============================
 
 ### BREAKING FEATURES
@@ -27,6 +27,7 @@ git2rdata 0.0.4 (2019-05-06)
   * Fix typo's in documentation, vignettes and README.
   * Add a ROpenSci review badge to the README.
   * The functions handle `root` containing regex expressions.
+  * Rework `vignette("workflow", package = "git2rdata")`
 
 git2rdata 0.0.3 (2019-03-12)
 ============================
