### 2019-05-14 da19785
update timings in efficiency vignette
