+++ NEWS.md
@@ -27,7 +27,8 @@ git2rdata 0.0.4 (2019-05-14)
   * Fix typo's in documentation, vignettes and README.
   * Add a ROpenSci review badge to the README.
   * The functions handle `root` containing regex expressions.
-  * Rework `vignette("workflow", package = "git2rdata")`
+  * Rework `vignette("workflow", package = "git2rdata")`.
+  * Update timings in `vignette("efficiency", package = "git2rdata")`
 
 git2rdata 0.0.3 (2019-03-12)
 ============================
