### 2019-05-15 04dec23
README mentions a maximal file size
