### 2019-05-15 0e2d96d
minor changes to the plain text vignette
