### 2019-05-15 f465c58
use vignette() in documentation instead of a hard link to the pkgdown website (which might change in the future)
