+++ R/recent_commit.R
@@ -9,7 +9,7 @@
 #' object in an analysis. Since it refers to the most recent change of this
 #' file, it remains unchanged by committing changes to other files. You can
 #' also use it to track if data got updated, requirering an analysis to
-#' be rerun.
+#' be rerun. See `vignette("workflow", package = "git2rdata")`.
 #' @inheritParams write_vc
 #' @param root The root of a project. Can be a file path or a `git-repository`
 #' @param data does `file` refers to a data object (TRUE) or to a file (FALSE).
