+++ man/push.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/reexport.R
 \name{push}
 \alias{push}
-\title{Reexported function from git2r}
+\title{Reexported function from `git2r``}
 \description{
 See \code{\link[git2r]{push}}
 }
