+++ man/relabel.Rd
@@ -29,10 +29,9 @@ invisible \code{NULL}
 Imagine the situation where we have a dataframe with a factor variable and we
 have stored it with \code{write_vc(optimize = TRUE)}. The raw data file contains
 the factor indices and the metadata contains the link between the factor
-index and the corresponding label. See the vignette on
-\href{https://inbo.github.io/git2rdata/articles/version_control.html#handling-factors-optimized}{version contol}
-(\code{vignette("version_control", package = "git2rdata")}) In such a case, relabeling a factor can be
-fast and lightweight by updating the metadata.
+index and the corresponding label. See
+\code{vignette("version_control", package = "git2rdata")}. In such a case,
+relabeling a factor can be fast and lightweight by updating the metadata.
 }
 \examples{
 
