### 2019-05-15 d9d078d
update lifecyle badges
