### 2019-05-15 c570c0a
use vignette() to link to other vignettes
