### 2019-05-15 2d1346f
require R 3.5.0 because vignette data was created using R 3.6.0 which are not compativle with R < 3.5.0
