+++ DESCRIPTION
@@ -14,7 +14,7 @@ Authors@R: c(
     "Research Institute for Nature and Forest",
     role = c("cph", "fnd"), email = "info@inbo.be"))
 Description: Make versioning of data.frame easy and efficient using git repositories.
-Depends: R (>= 3.4.0)
+Depends: R (>= 3.5.0)
 Imports:
   assertthat,
   git2r (>= 0.23.0),
