### 2019-05-15 7e678e3
tweak helpfile of reexported functions
