+++ R/reexport.R
@@ -1,6 +1,6 @@
-#' Reexported function from `git2r``
+#' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{repository}} in `git2r`
+#' See \code{\link[git2r]{repository}} in `git2r`.
 #' @name repository
 #' @importFrom git2r repository
 #' @family version_control
@@ -9,7 +9,7 @@ NULL
 
 #' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{status}} in `git2r`
+#' See \code{\link[git2r]{status}} in `git2r`.
 #' @name status
 #' @importFrom git2r status
 #' @family version_control
@@ -18,7 +18,7 @@ NULL
 
 #' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{commit}} in `git2r`
+#' See \code{\link[git2r]{commit}} in `git2r`.
 #' @name commit
 #' @importFrom git2r commit
 #' @family version_control
@@ -27,16 +27,16 @@ NULL
 
 #' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{pull}} in `git2r`
+#' See \code{\link[git2r]{pull}} in `git2r`.
 #' @name pull
 #' @importFrom git2r pull
 #' @family version_control
 #' @export
 NULL
 
-#' Reexported function from `git2r``
+#' Reexported function from `git2r`
 #'
-#' See \code{\link[git2r]{push}}
+#' See \code{\link[git2r]{push}} in  `git2r`.
 #' @name push
 #' @importFrom git2r push
 #' @family version_control
