+++ man/commit.Rd
@@ -4,7 +4,7 @@
 \alias{commit}
 \title{Reexported function from \code{git2r}}
 \description{
-See \code{\link[git2r]{commit}} in \code{git2r}
+See \code{\link[git2r]{commit}} in \code{git2r}.
 }
 \seealso{
 Other version_control: \code{\link{pull}},
