### 2019-05-15 4798bc2
use Title Case in documentation and vignettes
