+++ R/is_git2rdata.R
@@ -1,4 +1,4 @@
-#' Check whether a git2rdata object is valid.
+#' Check Whether a Git2rdata Object is Valid.
 #'
 #' A valid git2rdata object has valid metadata. The data hash must match the
 #' data hash stored in the metadata.
