### 2019-05-15 5b740f6
meta.factor() returns factor labels with UTF8 encoding. Fixes #34
