### 2019-05-15 3b16e8b
add internal function metadata_hash()
