### 2019-05-15 2b9726f
bugfix in recent_commit() on Windows
