+++ R/recent_commit.R
@@ -86,15 +86,18 @@ recent_commit.default <- function(file, root, data = FALSE) {
 
 #' @export
 #' @importFrom assertthat assert_that is.string is.flag noNA
-#' @importFrom git2r odb_blobs last_commit
+#' @importFrom git2r odb_blobs last_commit workdir
 recent_commit.git_repository <- function(file, root, data = FALSE) {
   assert_that(is.string(file), is.flag(data), noNA(data))
 
+  path <- unique(dirname(file))
+  if (path == ".") {
+    path <- ""
+  }
   if (data) {
-    file <- clean_data_path(root = root, file, normalize = FALSE)
+    file <- clean_data_path(root = workdir(root), file, normalize = FALSE)
   }
   name <- basename(file)
-  path <- gsub("^\\./?", "", unique(dirname(file)))
   blobs <- odb_blobs(root)
   blobs <- blobs[blobs$path == path & blobs$name %in% name, ]
   blobs <- blobs[blobs$when <= as.data.frame(last_commit(root))$when, ]
