### 2019-05-15 75f2c0c
add unit test to optain full coverage
