### 2019-05-16 f4aacbf
calculate relative git repository sizes from the data
