### 2019-05-16 832914b
add a section "Why to use git2rdata" to the README
