### 2019-05-16 cfa3d02
improve documentation of meta()
