### 2019-05-16 08c2bec
explicitly mention that read_vc() is an S3 method
