### 2019-05-16 d8680a7
update NEWS.md
