### 2019-05-16 #263 @ThierryO
Version 0.0.4 is ready for review.

## Changes as listed in NEWS.md

### BREAKING FEATURES

  * `write_vc()` stores the `git2rdata` version number to the metadata. Use `upgrade_data()` to update existing data.

### NEW FEATURES

  * `read_vc()` checks the meta data hash. A mismatch results in an error.
  * The meta data gains a data hash. A mismatch throws a warning when reading the object. This tolerates updating the data by other software, while informing the user that such change occurred.
  * `is_git2rmeta()` validates metadata.
  * `list_data()` lists files with valid metadata.
  * `rm_data()` and `prune_meta()` remove files with valid metadata. Other files are untouched.
  *  Files with invalid metadata yield a warning with `list_data()`, `rm_data()` and `prune_meta()`.

### Bugfixes

  * `write_vc()` and `relabel()` handle empty strings (`''`) in characters and factors (#24).
  * `read_vc()` no longer treats `#` as a comment character.
  * `read_vc()` handles non ASCII characters on Windows.

### Other changes

  * Use a faster algorithm to detect duplicates (suggestion by @brodieG).
  * Improve documentation.
  * Fix typo's in documentation, vignettes and README.
  * Add a ROpenSci review badge to the README.
  * The README mentions on upper bound on the size of dataframes.
  * Set lifecycle to "maturing" and repo status to "active".
  * The functions handle `root` containing regex expressions.
  * Rework `vignette("workflow", package = "git2rdata")`.
  * Update timings in `vignette("efficiency", package = "git2rdata")`
  * Minor tweaks in `vignette("plain_text", package = "git2rdata")`

## Changes not explicitly mentioned in NEWS.md

- README get's a section "Why use git2rdata".
- Titles are now in Title Case.
- POSIXct is indeed stored as numeric. This is corrected in the vignette.
- The timezone is set correctly. `print.data.frame()` doesn't show the timezone.
- Examples and unit tests clean up files.
- Reexported functions explicitly mention their originating package.
- The `root` is removed from a file path without using regex.
- `if (isTRUE(x)) {` is replaced by `is.flag(x); if (x) {`.
- `tz` argument is removed from `format()` in `meta.Date()`.
- Suggested patches were applied.
- `order()` uses `unname(x[sorting])`.
- The README.md starts by encouraging the user to visit the pkgdown version which contains a rendered version of the vignettes.
- Examples get more comments.

## Suggestions not implemented

### File structure

- The file structure remains as is, keeping it as open as possible by design.
- `vignette("workflow", package = "git2rdata")` gained a section "Structuring Git2rdata Objects Within a Project".
- Any change to the data something else that `git2rdata` will result in a warning. Since we recommend to use version control, the user can revert unwanted changes using version control.

### Warning for unstaged data

- `git2rdata` helps in making the analysis reproducible and traceable. It doesn't enforce it. That would be hard to do with forcing the user in a particular workflow. `vignette("workflow", package = "git2rdata")` suggests two possible workflows.

### Checking that the dispatching parameter actually is what it is supposed to be.

- That is what S4 does... Moreover, only the generics are exported. So the user must use the triple colon notation to use the functions directly (e.g. `git2rdata:::meta.character(1:10)`.

### Storing timezone

- I've tried that and it proved to be hard to get it 100% right.

### Pattern in `rm_data()` and `prune_meta()`

- The functions are intended to clean an entire working directory while (temporary) keeping the metadata. A pattern is not relevant is such a case. See `vignette("workflow", package = "git2rdata")`.

### Automatic naming of file

- The user needs the file argument to structure the data in the repository.
- Similar functions from `base`, `readr` and `data.table` all require the user to specify the filename.

### CONTRIBUTING file in the root

- It was moved to `.github` as suggested in inbo/git2rdata#10.

### message() instead of warning() when `strict = FALSE`

- The things triggering the warnings indicate potentially large diffs when using version control (which is a main reason for using git2rdata). Large diffs is what we try to avoid as much as possible. A warning seems more appropriate than a message.

### Different examples for `read_vc()` and `write_vc()`

- We use Roxygen templates for the examples of related functions as that is easier to maintain. The templates are available in `man-roxygen`.
