### 2019-05-20 #263 @brodieG
From a quick skim this looks good.  I'll review more thoroughly sometime next week and report back.
