### 2019-05-21 a43c35c
upgrade_data() preserves the order of the metadata
