### 2019-05-21 9712652
add inbo logo to sticker
