### 2019-05-31 #263 @brodieG
Looks good to me.  I did not test all the new functionality, but did confirm that the accidental yml removal which was the blocking issue is fixed.  It also does look like all the author-accepted suggestions are indeed in the new version, and I have no concerns with the author-rejected suggestions.  This was all tested against v0.0.4 tag.

Nice job finding corner cases in the new duplicate function.

Some minor comments, none of which gets in the way of approval:

* Re size: it would be useful to give an idea of how much repository grows with
  each additional commit to a file.  Even though a file may be 100MB, it doesn't
  mean you only get 10 commits before you're done.  As a user, I would want to
  know, "If I have a table of file x, how many times can I modify it and commit
  it before it becomes a problem".  Something like "For one 100MB file with 10%
  of rows changing between each commit, you would need X commits to reach 1GB
  repo size".  Obviously YMMV, but as a user I want to have a rough idea whether
  it is plausible that this could work for my use case (sum(file.size *
  commit.count)).
* Consider providing links to the HTML vignettes in README so that they may be
  read without installing the package.
* Typo in prune.R: "inclusing untracked files" -> "including"

@noamross next steps?
