### 2019-06-02 #263 @noamross
Approved! Thanks @ThierryO  for submitting and @brodieG and @jtr13  for your reviews! I know this has been a long road, but it's been very thorough and we're glad to have this useful and well-vetted package in our collection.  @ThierryO, should you choose to address the remaining small comments, you may, but we are good to go.

To-dos:
- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do and can add your co-authors.
- [ ] Change any links in badges for CI and coverage to point to the `ropensci` URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`.

In our developer's guide, [this chapter](https://ropensci.github.io/dev_guide/collaboration.html) starts the 3rd section that's about guidance for maintenance _after_ acceptance.

Welcome aboard! We'd love to host a blog post about your package - either a [short introduction to it with one example](https://ropensci.org/tech-notes/) or [a longer post with some narrative about its development or something you learned, and an example of its use](https://ropensci.org/blog/). If you are interested, review [the instructions](https://github.com/ropensci/roweb2#contributing-a-blog-post), and tag @stefaniebutland in your reply. She will get in touch about timing and can answer any questions.
