### 2019-06-04 c82d4bd
remove the obsolete sticker folder
