### 2019-06-04 a57b6de
add attribution to @peterdesmet for the new logo
