+++ NEWS.md
@@ -1,3 +1,8 @@
+git2rdata 0.1 (2019-06-04)
+============================
+
+  * Use new logo (@peterdesmet, #37).
+
 git2rdata 0.0.5 (2019-05-21)
 ============================
 
