### 2019-06-04 7533501
update link to reflect the transfer to rOpenSci
