+++ README.md
@@ -8,16 +8,16 @@
 [![minimal R version](https://img.shields.io/badge/R%3E%3D-3.5.0-6666ff.svg)](https://cran.r-project.org/)
 [![DOI](https://zenodo.org/badge/147685405.svg)](https://zenodo.org/badge/latestdoi/147685405)
 
-[![Travis-CI Build Status](https://travis-ci.org/inbo/git2rdata.svg?branch=master)](https://travis-ci.org/inbo/git2rdata)
-[![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/a3idhi9f6ls9xu8r/branch/master?svg=true)](https://ci.appveyor.com/project/ThierryO/git2rdata/branch/master)
-[![codecov](https://codecov.io/gh/inbo/git2rdata/branch/master/graph/badge.svg)](https://codecov.io/gh/inbo/git2rdata)
+[![Travis-CI Build Status](https://travis-ci.org/ropensci/git2rdata.svg?branch=master)](https://travis-ci.org/ropensci/git2rdata)
+[![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/cfbjb835fqb3dc7m/branch/master?svg=true)](https://ci.appveyor.com/project/ThierryO/git2rdata-n60yg/branch/master)
+[![codecov](https://codecov.io/gh/ropensci/git2rdata/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/git2rdata)
 
-![GitHub forks](https://img.shields.io/github/forks/inbo/git2rdata.svg?style=social)
-![GitHub stars](https://img.shields.io/github/stars/inbo/git2rdata.svg?style=social)
-![GitHub code size in bytes](https://img.shields.io/github/languages/code-size/inbo/git2rdata.svg)
-![GitHub repo size](https://img.shields.io/github/repo-size/inbo/git2rdata.svg)
+![GitHub forks](https://img.shields.io/github/forks/ropensci/git2rdata.svg?style=social)
+![GitHub stars](https://img.shields.io/github/stars/ropensci/git2rdata.svg?style=social)
+![GitHub code size in bytes](https://img.shields.io/github/languages/code-size/ropensci/git2rdata.svg)
+![GitHub repo size](https://img.shields.io/github/repo-size/ropensci/git2rdata.svg)
 
-<p style="display:none">Please visit the git2rdata website at https://inbo.github.io/git2rdata/. The vignette code on the website link to a rendered version of the vignette. Functions have a link to their help file.</p>
+<p style="display:none">**Please visit the git2rdata website at https://ropensci.github.io/git2rdata/. The vignette code on the website link to a rendered version of the vignette. Functions have a link to their help file.**</p>
 
 ## Rationale
 
@@ -54,13 +54,13 @@ Install the development version
 
 # install with vignettes (recommended)
 remotes::install_github(
-  "inbo/git2rdata", 
+  "ropensci/git2rdata", 
   build = TRUE, 
   dependencies = TRUE, 
   build_opts = c("--no-resave-data", "--no-manual")
 )
 # install without vignettes
-remotes::install_github("inbo/git2rdata"))
+remotes::install_github("ropensci/git2rdata"))
 ```
 
 ## Usage in a Nutshell
@@ -107,7 +107,7 @@ Please use the output of `citation("git2rdata")`
 - `testthat`: R scripts with unit tests using the [testthat](http://testthat.r-lib.org/) framework
 - `vignettes`: source code for the vignettes describing the package
 - `man-roxygen`: templates for documentation in Roxygen format
-- `pkgdown`: additional source files for the `git2rdata` [website](https://inbo.github.io/git2rdata/)
[…]
