### 2019-06-04 866105a
add an indication of the maximum number of commits
