+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: git2rdata
 Title: Store and Retrieve Data.frames in a Git Repository
-Version: 0.0.5
+Version: 0.1
 Authors@R: c(
   person(
     "Thierry", "Onkelinx", role = c("aut", "cre"), 
