### 2019-06-04 #263 @brodieG
I did miss the message about the pkgdown site, or didn't make the connection in the next section.  It's probably fine as it is.  FWIW that link is broken right now.
