### 2019-06-07 0b47c0d
Update badge URLs
