### 2019-06-07 e819fe8
Preempt URL redirects
