### 2019-06-13 dc22e46
try PKGTYPE win.binary to fix AppVeyor error
