### 2019-06-14 8353494
add cran-comments
