+++ cran-comments.md
@@ -1,0 +1,17 @@
+## Test environments
+* local
+    * ubuntu 18.04, R 3.6.0
+* travis-ci
+    * ubuntu 14.04, oldrel, release and devel
+* AppVeyor 
+    * Windows Server 2012, R 3.6.0 Patched
+* r-hub
+    * Windows Server 2008 R2 SP1, R-devel, 32/64 bit
+    * Ubuntu Linux 16.04 LTS, R-release, GCC
+    * Fedora Linux, R-devel, clang, gfortran
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
