### 2019-06-14 41c0fad
use absolute URL's in NEWS.md and README.md
