### 2019-06-14 6c0fa91
fix spelling errors
