+++ R/read_vc.R
@@ -3,7 +3,7 @@
 #' @description
 #' `read_vc()` handles git2rdata objects stored by `write_vc()`. It reads and
 #' verifies the metadata file (`.yml`). Then it reads and verifies the raw data.
-#' The last step is backtransforming any transformation done by `meta()` to
+#' The last step is back-transforming any transformation done by `meta()` to
 #' return the `data.frame` as stored by `write_vc()`.
 #'
 #' `read_vc()` is an S3 generic on `root` which currently handles `"character"`
