+++ README.md
@@ -132,4 +132,4 @@ git2rdata
 
 Contributions to `git2rdata` are welcome. Please read our [Contributing guidelines](https://github.com/ropensci/git2rdata/blob/master/.github/CONTRIBUTING.md) first. The `git2rdata` project is released with a [Contributor Code of Conduct](https://github.com/ropensci/git2rdata/blob/master/.github/CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
 
-[![ropensci_footer](http://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
+[![rOpenSci footer](http://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
