### 2019-06-14 dd30d09
fix spelling errors
