+++ R/recent_commit.R
@@ -8,7 +8,7 @@
 #' Use this information to document the current version of file or git2rdata
 #' object in an analysis. Since it refers to the most recent change of this
 #' file, it remains unchanged by committing changes to other files. You can
-#' also use it to track if data got updated, requirering an analysis to
+#' also use it to track if data got updated, requiring an analysis to
 #' be rerun. See `vignette("workflow", package = "git2rdata")`.
 #' @inheritParams write_vc
 #' @param root The root of a project. Can be a file path or a `git-repository`.
