### 2019-06-15 7016548
make unit test for subsecond commits more robust
