### 2019-06-15 761b082
travis builds the package on trusty, xenial and os x
