### 2019-06-15 0ccc753
update cran-comments
