### 2019-06-17 9dd4a0f
add CRAN and Rdocumentation badges to README
