### 2019-06-26 27c3793
use svg logo instead of png
