### 2019-06-26 5c97075
use svg logo instead of png
