+++ man/git2rdata-package.Rd
@@ -6,8 +6,6 @@
 \alias{git2rdata-package}
 \title{git2rdata: Store and Retrieve Data.frames in a Git Repository}
 \description{
-\if{html}{\figure{logo.png}{options: align='right'}}
-
 Make versioning of data.frame easy and efficient using git repositories.
 }
 \seealso{
