### 2019-06-26 8452019
improve warning() in case of problematic changes in the data
