### 2019-07-03 ed318d9
allow older but compatible versions of metadata
solves #44
