+++ NEWS.md
@@ -1,3 +1,10 @@
+git2rdata 0.1.0.9000 (2019-07-03)
+=================================
+
+  * Only require `upgrade_data()` for data written with versions prior to 0.0.5 (#44)
+  * Improve warnings() and error()
+  * Use vector version of logo
+
 git2rdata 0.1 (2019-06-04)
 ============================
 
