### 2019-08-13 670184f
improve issue template
fixes #48
