+++ .github/issue_template.md
@@ -1,9 +1,12 @@
 <!-- IF THIS INVOLVES AUTHENTICATION: DO NOT SHARE YOUR USERNAME/PASSWORD, OR API KEYS/TOKENS IN THIS ISSUE - MOST LIKELY THE MAINTAINER WILL HAVE THEIR OWN EQUIVALENT KEY -->
 
-<!-- If this issue relates to usage of the package, whether a question, bug or similar, along with your query, please paste your devtools::session_info() or sessionInfo() into the code block below, AND include a reproducible example (consider using a "reprex" https://cran.rstudio.com/web/packages/reprex/) If not, delete all this and proceed :) -->
+<!-- If this issue relates to usage of the package, whether a question, bug or similar, along with your query, please paste your sessioninfo::session_info() or sessionInfo() into the code block below, AND include a reproducible example (consider using a "reprex" https://cran.rstudio.com/web/packages/reprex/) If not, delete all this and proceed :) -->
+
+```r
+# replace this with your reproducible example
+```
 
 <details> <summary><strong>Session Info</strong></summary>
-
 ```r
 
 ```
