### 2019-08-13 cbceda4
use new style templates
