### 2019-08-13 f449659
clean_data_path() checks that file is contained within root
fixes #50
