### 2019-08-13 bd5bbce
clean_data_path() checks that file is contained within root
fixes #50
