+++ R/clean_data_path.R
@@ -11,6 +11,10 @@
 clean_data_path <- function(root, file, normalize = TRUE) {
   assert_that(is.flag(normalize), noNA(normalize))
   dir_name <- dirname(file)
+  if (length(grep("\\.\\.", dir_name))) {
+    stop("file should not contain '..'")
+  }
+
   file <- gsub("\\..*$", "", basename(file))
   if (dir_name == ".") {
     path <- file.path(root, file)
