+++ man/is_git2rdata.Rd
@@ -9,7 +9,8 @@ is_git2rdata(file, root = ".", message = c("none", "warning", "error"))
 \arguments{
 \item{file}{the name of the git2rdata object. Git2rdata objects cannot
 have dots in their name. The name may include a relative path. \code{file} is a
-path relative to the \code{root}.}
+path relative to the \code{root}.
+Note that \code{file} must point to a location within \code{root}.}
 
 \item{root}{The root of a project. Can be a file path or a \code{git-repository}.
 Defaults to the current working directory (\code{"."}).}
