+++ R/write_vc.R
@@ -28,7 +28,7 @@
 #' @export
 #' @family storage
 #' @template example-io
-#' @note `..generic` is a reserved name for the metadata and cannot be used as
+#' @note `..generic` is a reserved name for the metadata and is a forbidden
 #' column name in a `data.frame`.
 write_vc <- function(
   x, file, root = ".", sorting, strict = TRUE, optimize = TRUE, na = "NA",
